% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{AIC,CalmrFit-method}
\alias{AIC,CalmrFit-method}
\title{AIC of CalmrFit}
\usage{
\S4method{AIC}{CalmrFit}(object, ..., k = 2)
}
\arguments{
\item{object}{An object of class \code{\link{CalmrFit-class}}.}

\item{...}{Unused}

\item{k}{Numeric. Penalty term (default = 2)}
}
\value{
A numeric
}
\description{
Returns the Akaike Information Criterion of a model fit via calmr::fit_model
}
\details{
The AIC is defined as \code{2*k - 2*-NLL}, where k is a penalty term and NLL is the negative log likelihood of the model.
}
