% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Run the calm shiny application}
\usage{
app(browser = TRUE, install_pkg = FALSE)
}
\arguments{
\item{browser}{If TRUE, the app is launched in a browser.
calm.app package from github. Defaults to FALSE.}
}
\description{
Run the calm shiny application
}
\note{
The shiny app is distributed separately from this package.
}
