% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{BIC,CalmrFit-method}
\alias{BIC,CalmrFit-method}
\title{BIC of CalmrFit}
\usage{
\S4method{BIC}{CalmrFit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{CalmrFit-class}}.}

\item{...}{Unused}
}
\value{
A numeric
}
\description{
Returns the Bayesian Information Criterion of a model fit via calmr::fit_model
}
\details{
The AIC is defined as \code{k*log(n) - 2*-NLL}, where k is the number of parameters in the model and n is the number of observations
}
