% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Build and return a model}
\usage{
build_model(model_name = NULL, env = parent.frame())
}
\arguments{
\item{model_name}{A string with a specific model. Currently supports one of \code{c("HDI2020", "HD2022")}.}
}
\description{
Build and return a model
}
