% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{calm_model_plot}
\alias{calm_model_plot}
\title{Create a plot with calm data}
\usage{
calm_model_plot(dat, type)
}
\arguments{
\item{dat}{A data.frame-like with data to use in the plot}

\item{type}{A character specifying the type of plot}
}
\value{
A ggplot object
}
\description{
Create a plot with calm data
}
\note{
You should probably be getting plots via
the \code{plot} method for CalmExperiments.
}
