% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calm_verbosity.R
\name{calm_verbosity}
\alias{calm_verbosity}
\title{Set verbosity options for calm}
\usage{
calm_verbosity(verbose)
}
\arguments{
\item{verbose}{A logical}
}
\description{
Whether to show verbosity messages and progress bars
}
\note{
Progress bars are handled by the progressr package.
}
