% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design.R
\docType{class}
\name{CalmDesign-class}
\alias{CalmDesign-class}
\title{S4 class for Calm designs}
\description{
S4 class for Calm designs
}
\section{Slots}{

\describe{
\item{\code{design}:}{A list containing design information.}
\item{\code{mapping}:}{A list containing the object mapping.}
\item{\code{raw_design}:}{The original data.frame.}
\item{\code{augmented}:}{Whether the object has been augmented.}
}
}

