% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design.R, R/anccr_helpers.R
\name{augment,CalmDesign-method}
\alias{augment,CalmDesign-method}
\alias{augment}
\alias{.anccrize_design}
\alias{show,CalmDesign-method}
\alias{mapping}
\alias{mapping,CalmDesign-method}
\alias{trials}
\alias{trials,CalmDesign-method}
\title{Augment CalmDesign}
\usage{
\S4method{augment}{CalmDesign}(object, model, ...)

.anccrize_design(object)

\S4method{show}{CalmDesign}(object)

mapping(object)

\S4method{mapping}{CalmDesign}(object)

trials(object)

\S4method{trials}{CalmDesign}(object)
}
\arguments{
\item{object}{A CalmDesign, as returned by parse_design}

\item{model}{A modelname string. One of supported_models()}

\item{...}{Additional parameters depending on the model.}
}
\description{
Methods mapping, and trials, are extractor functions.
}
