% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_fit.R
\name{show,CalmFit-method}
\alias{show,CalmFit-method}
\alias{predict,CalmFit-method}
\alias{NLL}
\alias{NLL,CalmFit-method}
\alias{AIC,CalmFit-method}
\alias{BIC,CalmFit-method}
\title{CalmFit Methods}
\usage{
\S4method{show}{CalmFit}(object)

\S4method{predict}{CalmFit}(object, type = "response", ...)

NLL(object, ...)

\S4method{NLL}{CalmFit}(object)

\S4method{AIC}{CalmFit}(object, k = 2)

\S4method{BIC}{CalmFit}(object)
}
\arguments{
\item{object}{A CalmFit}

\item{type}{A string specifying the type of prediction to generate}

\item{...}{Additional arguments}

\item{k}{Penalty term}
}
\description{
CalmFit Methods

CalmFit methods
}
\details{
The AIC is defined as \code{2*k - 2*-NLL}, where k is a penalty
term and NLL is the negative log likelihood of the model.

The BIC is defined as \code{k*log(n) - 2*-NLL}, where k is the number
of parameters in the model and n is the number of observations
}
