% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\docType{class}
\name{CalmModel-class}
\alias{CalmModel-class}
\title{S4 class for Calm Models}
\description{
S4 class for Calm Models
}
\note{
Currently unused.
}
\section{Slots}{

\describe{
\item{model}{A model name string}
\item{model_fn}{The model function}
\item{state}{A list with the model internal state}
\item{parameters}{A list with the model with model parameters}
}
}

