% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmr_function.R
\docType{package}
\name{CaLMR-package}
\alias{CaLMR}
\alias{CaLMR-package}
\title{Package imports}
\description{
R functions for conducting MR analysis for Causal analysis of Latent exposures using Mendelian Randomization (CaLMR).
}
\author{
\strong{Maintainer}: Yue Yu \email{yy4778@princeton.edu}

Other contributors:
\itemize{
  \item Andrew Lakkis \email{andrew.lakkis@pennmedicine.upenn.edu} [contributor]
}

}
\keyword{internal}
