% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmr_app.R
\name{calmr_app}
\alias{calmr_app}
\title{Run the calmr GUI}
\usage{
calmr_app(browser = getOption("shiny.launch.browser", interactive()))
}
\arguments{
\item{browser}{If TRUE, the app is launched in a browser.}
}
\description{
Starts a shiny application
}
