% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{calmr_model_plot}
\alias{calmr_model_plot}
\title{Create a plot with calmr data}
\usage{
calmr_model_plot(dat, type)
}
\arguments{
\item{dat}{An \code{data.table} containing aggregated
data from a \linkS4class{CalmrExperiment}}

\item{type}{A character specifying the type of plot.}
}
\value{
A 'ggplot' object.
}
\description{
Create a plot with calmr data
}
\note{
You should probably be getting plots via
the \code{\link[=plot]{plot()}} method for \linkS4class{CalmrExperiment}.
}
