% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CalmrComparison-class}
\alias{CalmrComparison-class}
\title{S4 class for Calmr Comparisons}
\description{
S4 class for Calmr Comparisons
}
\section{Slots}{

\describe{
\item{\code{results}:}{Data.frame. The unparsed model results}
\item{\code{models}:}{Character vector. The models being compared}
\item{\code{layers}:}{Character vector. The layers being compared}
\item{\code{model_layer_names}:}{Character vector. Combination of the two above}
}
}

