% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design.R
\docType{class}
\name{CalmrDesign-class}
\alias{CalmrDesign-class}
\title{S4 class for Calmr designs}
\description{
Inherits from the tbl class
}
\section{Slots}{

\describe{
\item{\code{design}:}{A tbl containing the object.}
\item{\code{mapping}:}{A list containing the object mapping.}
\item{\code{raw_design}:}{The original data.frame.}
\item{\code{augmented}:}{Whether the object has been augmented.}
}
}

