% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_class.R
\docType{class}
\name{CalmrDesign}
\alias{CalmrDesign}
\title{S4 class for Calmr designs}
\description{
Inherits from the tbl class
}
\section{Methods}{

\describe{
\item{\code{trials}:}{Prints trial information per group and phase}
}
}

