% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R
\name{show,CalmrExperiment-method}
\alias{show,CalmrExperiment-method}
\alias{design,CalmrExperiment-method}
\alias{design}
\alias{trials,CalmrExperiment-method}
\alias{parameters<-}
\alias{parameters}
\alias{parameters,CalmrExperiment-method}
\alias{parameters<-,CalmrExperiment-method}
\alias{experiences,CalmrExperiment-method}
\alias{experiences}
\alias{results,CalmrExperiment-method}
\alias{results}
\alias{raw_results,CalmrExperiment-method}
\alias{raw_results}
\alias{parsed_results,CalmrExperiment-method}
\alias{parsed_results}
\alias{length,CalmrExperiment-method}
\alias{parse}
\alias{parse,CalmrExperiment-method}
\alias{aggregate,CalmrExperiment-method}
\alias{plot,CalmrExperiment-method}
\alias{plot}
\alias{graph}
\alias{graph,CalmrExperiment-method}
\alias{graph,CalmrExperiment}
\title{CalmrExperiment methods}
\usage{
\S4method{show}{CalmrExperiment}(object)

\S4method{design}{CalmrExperiment}(x)

\S4method{trials}{CalmrExperiment}(object)

parameters(x) <- value

\S4method{parameters}{CalmrExperiment}(x)

\S4method{parameters}{CalmrExperiment}(x) <- value

\S4method{experiences}{CalmrExperiment}(x)

\S4method{results}{CalmrExperiment}(object)

\S4method{raw_results}{CalmrExperiment}(object)

\S4method{parsed_results}{CalmrExperiment}(object)

\S4method{length}{CalmrExperiment}(x)

parse(object)

\S4method{parse}{CalmrExperiment}(object)

\S4method{aggregate}{CalmrExperiment}(x, ...)

\S4method{plot}{CalmrExperiment}(x, type = NULL, ...)

graph(x, ...)

\S4method{graph}{CalmrExperiment}(x, ...)
}
\arguments{
\item{object}{A CalmrExperiment object}

\item{x}{A CalmrExperiment}

\item{value}{A list of parameter lists.}

\item{...}{Additional parameters passed to \code{graph_calmr_model}}

\item{type}{character vector specifying the types of plots to create.
See ??supported_plots. Defaults to NULL.}
}
\value{
A ggplot object
}
\description{
Creates plots (or plot) with aggregated results in CalmrExperiment
}
\note{
With type = NULL, all supported plots are returned.
}
