% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R
\docType{class}
\name{CalmrExperiment-class}
\alias{CalmrExperiment-class}
\title{S4 class for calmr experiments.}
\description{
S4 class for calmr experiments.
}
\section{Slots}{

\describe{
\item{\code{design}:}{A \linkS4class{CalmrDesign} object.}
\item{\code{groups}:}{A string specifying the groups in the design.}
\item{\code{model}:}{A string specifying the model used.}
\item{\code{parameters}:}{A list with the parameters used, per group.}
\item{\code{timings}:}{A list with the timings used in the design.}
\item{\code{experiences}:}{A list with the experiences for the model.}
\item{\code{results}:}{A list with aggregated results.}
\item{\code{models}:}{The models associated with the iteration.}
\item{\code{.groups}:}{Internal. The groups associated with the iteration.}
\item{\code{.iter}:}{Internal. The iteration number.}
\item{\code{.seed}:}{The seed used to generate the experiment.}
}
}

\seealso{
CalmrExperiment-methods
}
