% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_result.R
\docType{class}
\name{CalmrExperimentResult-class}
\alias{CalmrExperimentResult-class}
\title{S4 class for calmr experiment results}
\description{
S4 class for calmr experiment results
}
\section{Slots}{

\describe{
\item{aggregated_results}{A list of \code{data.table} objects
with aggregated results.}
\item{parsed_results}{A list containing \code{data.table} objects
with parsed results.}
\item{raw_results}{A list with raw model outputs.}
}
}

