% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_fit.R
\docType{class}
\name{CalmrFit-class}
\alias{CalmrFit-class}
\title{S4 class for calmr Fit}
\description{
S4 class for calmr Fit
}
\section{Slots}{

\describe{
\item{\code{nloglik}:}{Numeric. Negative log likelihood of the fit}
\item{\code{best_pars}:}{Numeric. Best fitting parameters}
\item{\code{model_pars}:}{Numeric. Parameters used in the model function}
\item{\code{link_pars}:}{Numeric. Parameters used in the link function}
\item{\code{data}:}{Numeric. Data used for fit}
\item{\code{model_function}:}{Function. Model function}
\item{\code{link_function}:}{Function. Link function}
\item{\code{ll_function}:}{Function. Objective function
(usually nloglikelihood)}
\item{\code{optimizer_options}:}{List. Options used for the optimizer}
\item{\code{extra_pars}:}{List. Extra parameters
passed to the fit call (...)}
}
}

\seealso{
CalmrFit-methods
}
