% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_result.R
\name{show,CalmrResult-method}
\alias{show,CalmrResult-method}
\title{Methods for CalmrResult}
\usage{
\S4method{show}{CalmrResult}(object)
}
\arguments{
\item{object}{A CalmrResult object}
}
\description{
Methods for CalmrResult
}
