% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CalmrResult}
\alias{CalmrResult}
\title{S4 class for  Calmr results}
\description{
S4 class for  Calmr results
}
\section{Slots}{

\describe{
\item{\code{results}:}{Data.frame or List. The unparsed model results}
\item{\code{parsed_results}:}{List. The parsed and aggregated model results}
\item{\code{is_parsed}:}{Logical. Whether the model results have been parsed}
}
}

