% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CalmrRSA-class}
\alias{CalmrRSA-class}
\title{S4 class for Calmr RSA}
\description{
S4 class for Calmr RSA
}
\section{Slots}{

\describe{
\item{\code{corr_mat}:}{Array. Correlation matrix}
\item{\code{distance_mats}:}{List. Distance matrices}
\item{\code{trials}:}{Character. Trial option used to calculate distance mats}
\item{\code{dist_method}:}{Character vector. Method to calculate distances}
\item{\code{corr_method}:}{Character vector. Method to calculate correlations}
}
}

