% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CalmrRSATest-class}
\alias{CalmrRSATest-class}
\title{S4 class for Calmr RSA Test}
\description{
S4 class for Calmr RSA Test
}
\section{Slots}{

\describe{
\item{\code{RSA}:}{Object of class CalmrRSA}
\item{\code{sig_mat}:}{Array. Correlation significance matrix}
\item{\code{lower_crit}:}{Array. Lower bound for significance}
\item{\code{upper_crit}:}{Array. Upper bound for significance}
\item{\code{n_samples}:}{Numeric. Number of bootstrap samples}
\item{\code{p}:}{Numeric. Alpha level for significance}
}
}

