% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rsa.R
\docType{class}
\name{CalmRSA-class}
\alias{CalmRSA-class}
\title{S4 class for Calm representational similarity analysis}
\description{
S4 class for Calm representational similarity analysis
}
\section{Slots}{

\describe{
\item{\code{corr_mat}:}{Array. Correlation matrix}
\item{\code{distances}:}{List. Pairwise distance matrices}
\item{\code{args}:}{List. Arguments used to create the object.}
\item{\code{test_data}:}{List.
Test data, populated after testing the object.}
}
}

