% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rsa.R
\name{show,CalmRSA-method}
\alias{show,CalmRSA-method}
\alias{test}
\alias{test,CalmRSA-method}
\alias{plot,CalmRSA-method}
\title{CalmRSA Methods}
\usage{
\S4method{show}{CalmRSA}(object)

test(object, n_samples = 1000, p = 0.95)

\S4method{test}{CalmRSA}(object, n_samples = 1000, p = 0.95)

\S4method{plot}{CalmRSA}(x, y, ...)
}
\arguments{
\item{object}{A CalmRSA object}

\item{n_samples}{The number of samples for the permutation test
(default = 1e3)}

\item{p}{The critical threshold level for the permutation test
(default = 0.95)}

\item{x}{A CalmRSA object to plot}

\item{y}{Unused.}

\item{...}{Extra parameters passed to the plot call}
}
\value{
A CalmRSA object with the test results
}
\description{
CalmRSA Methods

Test CalmRSA object via permutation test
}
