% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Run models given a set of parameters}
\usage{
compare_models(
  design,
  models,
  layers,
  parameters = NULL,
  options = get_exp_opts()
)
}
\arguments{
\item{design}{A design data.frame}

\item{models}{A character vector specifying the models to run.}

\item{layers}{A character vector of length M (the length of \code{models}),
specifying the layers to compare
specifying the layer in each model that will be used for layers.}

\item{parameters}{A list of length M,
each containing a list with model parameters,
as returned by `get_parameters``}

\item{options}{A list of options, as returned by get_exp_opts.}
}
\value{
A CalmrComparison object.f
}
\description{
Run models given a set of parameters
}
