% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmr_function.R
\name{convert_to_CaLMR_format}
\alias{convert_to_CaLMR_format}
\title{Convert merged GWAS data to CaLMR input format}
\usage{
convert_to_CaLMR_format(df, outcome_trait)
}
\arguments{
\item{df}{Data frame of merged SNP-level data (output of
\code{LD_clumped_SNPs}).}

\item{outcome_trait}{Character string identifying the outcome.}
}
\description{
Convert merged GWAS data to CaLMR input format
}
\keyword{internal}
