% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parsers.R
\name{.parse_experiment}
\alias{.parse_experiment}
\title{An assorment of functions to help parse models}
\usage{
.parse_experiment(experiment)
}
\description{
An assorment of functions to help parse models
}
