% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_helpers.R
\name{get_optimizer_opts}
\alias{get_optimizer_opts}
\alias{fit_predict}
\title{An assortment of functions to help fit HeiDI}
\usage{
get_optimizer_opts(optimizer, stim_names, family, adj = 1e-06)

fit_predict(fit, new_args = NULL, type = "response")
}
\arguments{
\item{optimizer}{A string specifying the optimizer to use. One from \code{c("optim")}}

\item{stim_names}{A string specifying the name of stimuli in the experiment}

\item{family}{A string specifying the family function to generate responses (and calculate MLE with). Currently supports \code{c("identity", "linear", "poisson")}}

\item{adj}{The adjustment factor for upper and lower bounds. Default is 1e-6}
}
\description{
An assortment of functions to help fit HeiDI
}
\note{
Whenever a family function other than the identity is used, the family-specific parameters will always be appended to the end of the relevant list entries.
}
