% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_support_functions.R
\name{gen_ss_weights}
\alias{gen_ss_weights}
\title{Model support functions}
\usage{
gen_ss_weights(stims, default_val = 0)
}
\description{
An assortment of functions to support models (mostly HeiDI)
}
