% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_graphs.R
\name{get_graph_opts}
\alias{get_graph_opts}
\title{Get options for calmr graph}
\usage{
get_graph_opts(graph_size = "small")
}
\arguments{
\item{size}{A character (one of "small" or "large")
to return default values for small or large graphs}
}
\value{
A list with graph options, to be passed to \code{ggnetwork::geom_nodes}
}
\description{
Get options for calmr graph
}
