% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heidi_opts.R
\name{get_heidi_opts}
\alias{get_heidi_opts}
\title{Get options for heidi model}
\usage{
get_heidi_opts(iterations = 1)
}
\arguments{
\item{iterations}{An integer denoting the number of iterations to run}
}
\description{
Get options for heidi model
}
\examples{
get_heidi_opts(iterations = 10)
}
