% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Return a model function}
\usage{
get_model(model_name)
}
\arguments{
\item{model_name}{A string specifying the model to be returned. One of \code{supported_models()}.}
}
\description{
Given a model string, returns a model function that can be used elsewhere in the package.
}
