% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_params.R
\name{get_model_params}
\alias{get_model_params}
\title{Obtain model parameters}
\usage{
get_model_params(design, model = NULL)
}
\arguments{
\item{design}{An data.frame containing the experimental design.}

\item{model}{A string specifying a model. One in \code{supported_models()}.}
}
\value{
A data.frame with model parameters, depending on \code{model}.
}
\description{
Obtain model parameters
}
\seealso{
\code{\link{supported_models}}
}
