% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Obtain model parameters}
\usage{
get_parameters(design, model = NULL)
}
\arguments{
\item{design}{An data.frame containing the experimental design.}

\item{model}{A string specifying a model. One in \code{supported_models()}.}
}
\value{
A list with model parameters depending on model
}
\description{
Obtain model parameters
}
\seealso{
\code{\link{supported_models}}
}
