% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Returns a data.frame with parameters for stimuli contained in design}
\usage{
get_params(design, default_par)
}
\arguments{
\item{design}{A tibble with the experimental design, as returned by parse_design}

\item{default_par}{A float between 0 and 1.}
}
\value{
A data.frame with columns Stimulus and Alpha, for the stimulus name and its salience, respectively.
}
\description{
Returns a data.frame with parameters for stimuli contained in design
}
\seealso{
parse_design
}
