% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeidiOpts.R
\name{getHeidiOpts}
\alias{getHeidiOpts}
\title{Get options for heidi model}
\usage{
getHeidiOpts(iterations = 1)
}
\arguments{
\item{iterations}{An integer denoting the number of iterations to run}
}
\description{
Get options for heidi model
}
\examples{
getHeidiOpts(iterations = 10)
}
