% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParams.R
\name{getParams}
\alias{getParams}
\title{Returns a data.frame with parameters for stimuli contained in design}
\usage{
getParams(design_list, default_par)
}
\arguments{
\item{design_list}{A list with the experimental design, as returned by parseDesign}

\item{default_par}{A float between 0 and 1.}
}
\value{
A data.frame with columns Stimulus and Alpha, for the stimulus name and its salience, respectively.
}
\description{
Returns a data.frame with parameters for stimuli contained in design
}
\seealso{
parseDesign
}
