% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{graph,CalmrExperiment-method}
\alias{graph,CalmrExperiment-method}
\title{Graph model associations}
\usage{
\S4method{graph}{CalmrExperiment}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{CalmrModel-class}} or \code{\link{CalmrExperiment-class}}.}

\item{...}{Additional parameters passed to the \code{\link{graph_weights}} function.}
}
\value{
A ggplot object
}
\description{
Creates a network graph of model associations
}
