% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heidi_app.R
\name{heidi_app}
\alias{heidi_app}
\title{Run the heidi GUI}
\usage{
heidi_app(browser = getOption("shiny.launch.browser", interactive()))
}
\arguments{
\item{browser}{If TRUE, the app is launched in a browser.}
}
\description{
Starts a shiny application
}
