% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heidi_plots.R
\name{heidi_plots}
\alias{heidi_plots}
\alias{makePlots}
\alias{plotWeights}
\alias{plotVs}
\alias{plotRs}
\title{Plotting functions for heidi models}
\usage{
makePlots(dat)

plotWeights(vals)

plotVs(vals)

plotRs(vals)
}
\arguments{
\item{dat}{A list containing data.frames with parsed weights, r values, and v values, as returned by parseHeidiResults.}

\item{vals}{A data.frame containing parsed values}
}
\value{
A list with ggplot objects
}
\description{
Plotting functions for heidi models
}
\seealso{
parseHeidiResults
}
