% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertions.R
\name{is_experiment}
\alias{is_experiment}
\title{Check if object is Calmr experiment}
\usage{
is_experiment(object)
}
\arguments{
\item{object}{The object to be checked}
}
\value{
A logical
}
\description{
Check if object is Calmr experiment
}
