% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAC1975.R
\name{MAC1975}
\alias{MAC1975}
\title{Train the MAC1975 model}
\usage{
MAC1975(v = NULL, parameters, experience, mapping, ...)
}
\arguments{
\item{v}{(optional) A named matrix of dimensions S,S;
where S is the number of stimuli.}

\item{parameters}{A list containing the model parameters,
as returned by get_parameters().}

\item{experience}{A data.frame specifying trials as rows,
as returned by \code{make_experiment}}

\item{mapping}{A named list specifying trial and stimulus mapping,
as returned by \code{make_experiment}}

\item{...}{Additional named arguments}
}
\value{
A list with raw results
}
\description{
Train the MAC1975 model
}
