% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_experiment.R
\name{make_experiment}
\alias{make_experiment}
\title{Make CalmrExperiment}
\usage{
make_experiment(
  design,
  parameters = NULL,
  model = NULL,
  options = get_exp_opts()
)
}
\arguments{
\item{design}{A design data.frame}

\item{parameters}{Parameters for a  model as
returned by \code{get_parameters}}

\item{model}{A string specifying the model name. One of \code{supported_models()}}

\item{options}{A list with options as returned by \code{get_exp_opts}}
}
\value{
A CalmrExperiment object
}
\description{
Makes a CalmrExperiment object containing
the arguments necessary to run models
}
\examples{
des <- data.frame(Group = "G1", P1 = "10A>(US)", R1 = TRUE)
ps <- get_parameters(des, model = "HD2022")
op <- get_exp_opts(iterations = 1)
make_experiment(design = des, parameters = ps, model = "HD2022", options = op)

}
\seealso{
\code{\link{parse_design}},
\code{\link{get_parameters}}, \code{\link{get_exp_opts}}
}
