% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_heidi_args.R
\name{make_heidi_args}
\alias{make_heidi_args}
\title{Make a tibble to fit a heidi model}
\usage{
make_heidi_args(design, pars, opts = get_heidi_opts())
}
\arguments{
\item{design}{A design tibble, as returned by \code{parse_design}}

\item{pars}{A data.frame containing parameters as returned by \code{get_params}}

\item{opts}{A list with options as returned by \code{get_heidi_opts}}
}
\value{
A tibble with the arguments required to run the model. Each row represents a group in the experimental design.
}
\description{
Make a tibble to fit a heidi model
}
\examples{
df <- data.frame(Group = c('Group 1', 'Group 2'), P1 = c('10AB(US)', '10A(US)'), R1 = c(TRUE, TRUE))
des <- parse_design(df)
ps <- get_params(des, 0.2)
make_heidi_args(design = des, pars = ps, opts = get_heidi_opts(iterations = 1))

}
\seealso{
\code{\link{parse_design}}, \code{\link{get_params}}, \code{\link{get_heidi_opts}}
}
