% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeidiArgs.R
\name{makeHeidiArgs}
\alias{makeHeidiArgs}
\title{Make a tibble to fit a heidi model}
\usage{
makeHeidiArgs(design, pars, opts)
}
\arguments{
\item{design}{A design list, as returned by parseDesign}

\item{pars}{A parameter data.frame as returned by getParams}

\item{opts}{A list with options}
}
\value{
A tibble with the arguments required to run the model. Each row represents a group in the experimental design.
}
\description{
Make a tibble to fit a heidi model
}
\examples{
df <- data.frame(Group = c('Group 1', 'Group 2'), P1 = c('10AB(US)', '10A(US)'), R1 = c(TRUE, TRUE))
des <- parseDesign(df)
ps <- getParams(des, 0.2)
makeHeidiArgs(design = des, pars = ps, opts = list(iterations = 1))

}
\seealso{
parseDesign, getParams
}
