% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_functions.R
\name{information_functions}
\alias{information_functions}
\alias{supported_models}
\alias{supported_optimizers}
\alias{supported_families}
\alias{supported_plots}
\alias{get_model}
\alias{model_outputs}
\title{Miscellaneous information functions}
\usage{
supported_models()

supported_optimizers()

supported_families()

supported_plots(model = NULL)

get_model(model)

model_outputs(model = NULL)
}
\arguments{
\item{model}{A string specifying a model. One from \code{supported_models()}}
}
\description{
\describe{
\item{\code{supported_models}}{Returns the models supported in the package.}
\item{\code{supported_plots, model_outputs}}{Returns the plots, outputs
supported by argument \code{model}.}
\item{\code{supported_families}}{Returns the families supported by
\code{\link{fit_model}}.}
\item{\code{supported_optimizers}}{returns the optimizers
supported by \code{\link{fit_model}}.}
\item{\code{supported_plots}}{returns the plots supported by
argument \code{model}.}
\item{\code{get_model}}{Returns the function specified by argument \code{model}.}
\item{\code{parameter_info}}{Returns parameter information for a
\code{model} (or all of them if model = NULL)}
\item{\code{}}{}
}
}
