% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info_funcs.R
\name{supported_models}
\alias{supported_models}
\alias{.get_model_parnames}
\title{Miscellaneous model functions
\code{supported_models} returns the models supported in the package.
\code{supported_plots} returns the plots supported by argument \code{model}.}
\usage{
supported_models()

.get_model_parnames(model)
}
\arguments{
\item{model}{A string specifying a model. One in \code{supported_models()}}
}
\description{
Miscellaneous model functions
\code{supported_models} returns the models supported in the package.
\code{supported_plots} returns the plots supported by argument \code{model}.
}
