% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{NLL}
\alias{NLL}
\title{NLL of CalmrFit}
\usage{
NLL(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{CalmrFit-class}}.}

\item{...}{Unused}
}
\value{
A numeric
}
\description{
Returns the negative log likelihood of a model fit via calmr::fit_model
}
