% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R
\name{parameters}
\alias{parameters}
\alias{parameters<-}
\title{Retrieve/set parameters from/in CalmrExperiment}
\usage{
parameters(x)

parameters(x) <- value
}
\arguments{
\item{x}{A CalmrExperiment object}

\item{value}{A list of parameters}
}
\description{
Retrieve/set parameters from/in CalmrExperiment
}
\note{
If passing a list of parameters (as returned by \code{get_parameters}),
the set method will overwrite all rows in the experiment.
Passing a list of lists will replace on a row-by-row basis.
}
