% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_design.R
\name{parse_design}
\alias{parse_design}
\title{Parse design data.frame}
\usage{
parse_design(df, model = NULL, ...)
}
\arguments{
\item{df}{A data.frame of dimensions Groups, 2*Phases+1}

\item{model}{An optional model to augment the design. See ??augment design}

\item{...}{Other arguments passed to augment_design}
}
\value{
A CalmrDesign object
}
\description{
Parse design data.frame
}
\note{
\itemize{
\item{
Each entry in even-numbered columns of df is
a string formatted as per trial_parser.
}
}
}
\examples{
df <- data.frame(
  Group = c("Group 1", "Group 2"),
  P1 = c("10AB(US)", "10A(US)"), R1 = c(TRUE, TRUE)
)
parse_design(df)
}
\seealso{
\code{\link{trial_parser}}
}
