% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseDesign.R
\name{parseDesign}
\alias{parseDesign}
\title{Parse each string in a data.frame}
\usage{
parseDesign(df)
}
\arguments{
\item{df}{A data.frame of dimensions Groups, 2*Phases+1}
}
\value{
A list of length nrow(df), the number of groups in the table.
}
\description{
Parse each string in a data.frame
}
\note{
\itemize{
\item{
Each entry in even-numbered columns of df is a string formatted as per trialParser.
}
\item{
Each element of the returned list list is itself a list of length (ncol(df)-1)/2, the number of phases
}
}
}
\examples{
df <- data.frame(Group = c('Group 1', 'Group 2'), P1 = c('10AB(US)', '10A(US)'), R1 = c(TRUE, TRUE))
parseDesign(df)
}
\seealso{
trialParser
}
