% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHeidiResults.R
\name{parseHeidiResults}
\alias{parseHeidiResults}
\title{Parse heidi results}
\usage{
parseHeidiResults(raw_results)
}
\arguments{
\item{raw_results}{A tibble containing a mod_data column, as returned by runHeidi}
}
\value{
A list containing parsed ws, vs, and rs.
}
\description{
Parse heidi results
}
\seealso{
runHeidi
}
