% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_graphs.R
\name{patch_graphs}
\alias{patch_graphs}
\title{Patch Calm graphs}
\usage{
patch_graphs(graphs, selection = names(graphs))
}
\arguments{
\item{graphs}{A list of named graphs, as returned by \code{calm::graph}}

\item{selection}{A character or numeric vector determining the plots to patch}
}
\description{
Convenience function to patch graphs with cowplot
}
