% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{patch_plots}
\alias{patch_plots}
\title{Patch Calmr plots}
\usage{
patch_plots(plots, selection = names(plots), plot_options = get_plot_opts())
}
\arguments{
\item{plots}{A list of named plots, as returned by \code{calmr::plot}}

\item{selection}{A character or numeric vector determining the plots to patch}

\item{plot_options}{A list of plot options as returned by \code{get_plot_opts}}
}
\description{
Convenience function to patch plots with \code{patchwork}
}
