% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pati}
\alias{pati}
\title{Rat responses fromPatittuci et al. 2016}
\format{
A tibble with the following variables:
\describe{
\item{subject}{subject identifier}
\item{rpert}{responses per trial}
\item{lever}{lever presented on the trial: L = left; R = right}
\item{response}{the response: lp = lever press; np = nose poke}
\item{block}{the 2-session block of training (1 to 8)}
...
}
}
\source{
{Patittuci et al. (2016). JEP:ALC}
}
\usage{
pati
}
\description{
A dataset containing the rat nose pokes and lever presses when they presented with lever presses associated with different appetitive stimuli.
}
\keyword{datasets}
