% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phase_parser.R
\name{phase_parser}
\alias{phase_parser}
\title{Parses a phase string}
\usage{
phase_parser(phase_string)
}
\arguments{
\item{phase_string}{A string}
}
\value{
A named list
}
\description{
Parses a phase string
}
