% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot,CalmrRSA-method}
\alias{plot,CalmrRSA-method}
\title{Plot RSA}
\usage{
\S4method{plot}{CalmrRSA}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{\link{CalmrRSA-class}}.}

\item{...}{Additional parameters passed to the plotting function.}
}
\value{
A ggplot object
}
\description{
Plot a correlogram from RSA results
}
