% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot,CalmrRSATest-method}
\alias{plot,CalmrRSATest-method}
\title{Plot RSA test}
\usage{
\S4method{plot}{CalmrRSATest}(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{\link{CalmrRSATest-class}}.}

\item{...}{Additional parameters passed to the plotting function.}
}
\value{
A ggplot object
}
\description{
Plot a correlogram from RSA test results
}
