% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.r
\name{plot,CalmrExperiment-method}
\alias{plot,CalmrExperiment-method}
\title{Plot CalmrExperiment}
\usage{
\S4method{plot}{CalmrExperiment}(x, type = NULL, y = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{CalmrExperiment-class}}.}

\item{type}{A string specifying the type of plot to create.
See ??supported_plots.}

\item{...}{Additional parameters passed to the plotting function.}
}
\value{
A ggplot object
}
\description{
Creates a plot depicting results from CalmrExperiment
}
