% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict,CalmrFit-method}
\alias{predict,CalmrFit-method}
\title{Predict from CalmrFit}
\usage{
\S4method{predict}{CalmrFit}(object, type = "response", ...)
}
\arguments{
\item{object}{An object of class \code{\link{CalmrFit-class}}.}

\item{type}{A string. If \code{response}, model responses are transformed via the link function.}

\item{...}{Additional parameters passed to the function (\code{object@model_function}).}
}
\description{
Obtain a prediction from CalmrFit
}
