% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_heidi.R
\name{quick_heidi}
\alias{quick_heidi}
\title{Run the HeiDI model (quickly)}
\usage{
quick_heidi(design_df, param_df = NULL, options = NULL, parse = TRUE)
}
\arguments{
\item{design_df}{A data.frame of dimensions G,2*P+1; where G is the number of groups and P is the number of phases.}

\item{param_df}{A data.frame of dimensions N,2; where N is the number of stimuli in the experimental design.}

\item{options}{A list of options, as returned by \code{get_heidi_opts}.}

\item{parse}{A logical specifying whether the results should be parsed via \code{parse_heidi_results}. Default = TRUE.}
}
\value{
A list with parsed results or a tibble with raw results
}
\description{
Runs the model with minimal parameters.
}
\seealso{
\code{\link{get_heidi_opts}}, \code{\link{parse_heidi_results}}
}
