% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAND.R
\name{RAND}
\alias{RAND}
\title{Train the RAND model}
\usage{
RAND(v = NULL, parameters, experience, mapping, ...)
}
\arguments{
\item{v}{(optional) A named matrix of dimensions S,S;
where S is the number of stimuli.}

\item{parameters}{A list containing the model parameters,
as returned by parameter_info().}

\item{experience}{A data.frame specifying trials as rows,
as returned by \code{make_experiment}}

\item{mapping}{A named list specifying trial and stimulus mapping,
as returned by \code{make_experiment}}
}
\value{
A list with raw results
}
\description{
Train the RAND model
}
