% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{results,CalmrExperiment-method}
\alias{results,CalmrExperiment-method}
\title{Extract aggregated results from CalmrExperiment}
\usage{
\S4method{results}{CalmrExperiment}(object)
}
\arguments{
\item{object}{An object of class \clode{\link{CalmrExperiment}}}
}
\value{
A tbl containing models (rows) and model outputs (columns)
}
\description{
Extract aggregated results from CalmrExperiment
}
