% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R
\name{rsa,CalmrExperiment-method}
\alias{rsa,CalmrExperiment-method}
\alias{rsa}
\title{Perform representational similarity analysis on CalmrExperiment}
\usage{
\S4method{rsa}{CalmrExperiment}(x, comparisons, test = FALSE, ...)
}
\arguments{
\item{x}{A tbl of m by o (models by outputs) with aggregated results.}

\item{comparisons}{A model-named list containing the model
outputs to compare.}

\item{test}{Whether to test the RSA via permutation test. Default = FALSE.}

\item{...}{Additional parameters passed to \code{stats::dist}
and \code{stats::cor}}
}
\value{
A CalmrRSA object
}
\description{
Perform representational similarity analysis on CalmrExperiment
}
\note{
The object returned by this function
can be later tested via its own \code{test} method.
}
\examples{
# Comparing the associations in three models
exp <- data.frame(
  Group = c("A", "B"),
  P1 = c("2(A)>(US)/1B>(US)", "1(A)>(US)/2B>(US)"),
  R1 = TRUE
)
exp <- parse_design(exp)
models <- c("HD2022", "RW1972", "PKH1982")
parameters <- sapply(models, get_parameters, design = exp)
options <- get_exp_opts()
exp_res <- compare_models(exp,
  models = models,
  parameters = parameters, options = options
)
comparisons <- list(
  "HD2022" = c("vs"),
  "RW1972" = c("vs"),
  "PKH1982" = c("eivs")
)
res <- rsa(exp_res, comparisons = comparisons)
test(res, n_samples = 100)
}
