% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Run experiment}
\usage{
run_experiment(x, parse = TRUE, ...)
}
\arguments{
\item{x}{A CalmrExperiment or design data.frame}

\item{parse}{A logical specifying whether the results
should be parsed. Default = TRUE.}

\item{...}{Arguments to make the experiment
(in case x is a design data.frame)
or arguments for the model call (e.g., debug)}
}
\value{
A CalmrExperiment with results.
}
\description{
Runs an experiment with minimal parameters.
}
\examples{
# Using a data.frame only
df <- data.frame(
  Group = c("True", "Pseudo"),
  P1 = c("10AB(US)/10AC", "5AB(US)/5AB/5AC(US)/5AC"),
  R1 = c(TRUE, TRUE),
  P2 = c("1A", "1A"),
  R2 = c(TRUE, TRUE)
)
run_experiment(df)
run_experiment(model = "MAC1975")

# Using a CalmrExperiment with modified parameters
df <- data.frame(
  Group = c("Weak", "Strong"),
  P1 = c("10AB(US)", "10AC(US)"),
  R1 = FALSE
)
pars <- get_parameters(model = "HD2022")
exp <- make_experiment(df, parameters = pars)
run_experiment(exp)
}
\seealso{
\code{\link{get_exp_opts}}, \code{\link{parse_experiment_results}}
}
