% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runHeidi.R
\name{runHeidi}
\alias{runHeidi}
\title{Run the HeiDI model}
\usage{
runHeidi(design_df, param_df = NULL, options = NULL, parse = T)
}
\arguments{
\item{design_df}{A data.frame of dimensions G,2*P+1; where G is the number of groups and P is the number of phases.}

\item{param_df}{A data.frame of dimensions N,2; where N is the number of stimuli in the experimental design.}

\item{options}{A list of options, as returned by getHeidiOpts.}

\item{parse}{A logical specifying whether the results should be parsed. Default = TRUE.}
}
\value{
A list with parsed results or a tibble with raw results
}
\description{
Runs the model given a data.frame containing the experimental design and parameters.
}
\note{
The specification of design_df should adhere to the following rules:
\itemize{
\item{
The first column should specify the Group names
}
\item{
The remaining columns should be paired, each pair denoting a single experimental phase.
The first containing trial strings (see below) and logicals denoting whether the trials should be randomized or not.
}
\item{
Trial strings use a specific syntax. Take the string "30AX(US)/10A", for example.
\itemize{
\item{
The "/" character is used to denote different trials. Above, the string specifies two trial types.
}
\item{
Parentheses are used to denote complex stimuli (or stimuli that need to be specifically named). Above, US is a named stimulus.
}
\item{
The numbers at the beginning of each trial type denote the number of trials given within a phase.
}}
}
}
}
\seealso{
getHeidiOpts, parseDesign, trainPavHEIDI, parseHeidiResults
}
