% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{Example summary statistics input of the function: sample_data}
\format{
A data frame with 37 rows and 14 variables:
\describe{
  \item{beta.B1}{summary coef for trait B1.}
  \item{se.B1}{summary se for trait B1.}
  \item{beta.B2}{summary coef for trait B2.}
  \item{se.B2}{summary se for trait B2.}
  \item{beta.B3}{summary coef for trait B3.}
  \item{se.B3}{summary se for trait B3.}
  \item{beta.B4}{summary coef for trait B4.}
  \item{se.B4}{summary se for trait B4.}
  \item{beta.B5}{summary coef for trait B5.}
  \item{se.B5}{summary se for trait B5.}
  \item{beta.B6}{summary coef for trait B6.}
  \item{se.B6}{summary se for trait B6.}
  \item{beta.Y}{summary coef for outcome Y.}
  \item{se.Y}{summary se for outcome Y.}
}
}
\source{
Simulated data for the package example.
}
\usage{
sample_data
}
\description{
This dataset contains simulated summary data for six observable traits (B1 to B6) and an outcome variable (Y).
}
\keyword{datasets}
