% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_calmr_palette.R
\name{set_calmr_palette}
\alias{set_calmr_palette}
\title{Get/set the colour/fill palette for plots}
\usage{
set_calmr_palette(palette = NULL)
}
\arguments{
\item{palette}{A string specifying the available palettes.
If NULL, returns available palettes.}
}
\value{
The old palette (invisibly) if palette is not NULL.
Otherwise, a character vector of available palettes.
}
\description{
Get/set the colour/fill palette for plots
}
\note{
Changes here do not affect the palette used in graphs.
}
