% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anccr_helpers.R
\name{set_reward_parameters}
\alias{set_reward_parameters}
\title{Set reward parameters for ANCCR model}
\usage{
set_reward_parameters(parameters, rewards = c("US"))
}
\arguments{
\item{parameters}{A list of parameters, as returned by \code{\link[=get_parameters]{get_parameters()}}}

\item{rewards}{A character vector specifying the reward stimuli.
Default = \code{c("US")}}
}
\value{
A list of parameters
}
\description{
Set reward parameters for ANCCR model
}
\note{
The default behaviour of \code{get_parameters} for the ANCCR model is to
set every reward-related parameter to its non-zero default value.
This function will set those parameters to zero for non-reward stimuli
}
