% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_helpers.R
\name{shiny_helpers}
\alias{shiny_helpers}
\alias{patchPlots}
\title{Some functions to help the shiny application.
None of these are exported to the user.}
\usage{
patchPlots(plots, selection, options)
}
\arguments{
\item{plots}{A list of plots}

\item{selection}{A character vector with the selected plots}

\item{options}{A character vector with all available plots}
}
\description{
Some functions to help the shiny application.
None of these are exported to the user.
}
