% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SM2007.R
\name{SM2007}
\alias{SM2007}
\title{Train the SM2007 model}
\usage{
SM2007(
  alphas,
  lambdas,
  omegas,
  rhos,
  gammas,
  kappas,
  C,
  V = NULL,
  experience,
  mapping
)
}
\arguments{
\item{alphas}{A named vector with stimulus saliences.}

\item{lambdas}{A named vector with learning asymptotes.}

\item{omegas}{A named vector with weakening rates (extinction).}

\item{rhos}{A named vector with weights for activation of present stimuli.}

\item{gammas}{A named vector with weights for contribution of each stimulus in comparison processes.}

\item{kappas}{A named vector with stimulus-specific rates for operator switch.}

\item{C}{An integer specifying the depth of the comparison process. Defaults to 1.}

\item{V}{(optional) A named matrix of dimensions S,S; where S is the number of stimuli.}

\item{experience}{A data.frame specifying trials as rows, as returned by \code{make_model_args}}

\item{mapping}{A named list specifying trial and stimulus mapping, as returned by \code{make_model_args}}
}
\value{
A list with
\itemize{
\item{vs, Array of dimensions P,S,S; where P is the number of trials used to train the model and S is the number of stimuli involved in the experiment. Contains stimulus association strengths.}
\item{acts, As above but of dimensions P, S, S, C, containing relative activations on the n-th order (as determined by argument C).}
\item{other, Carryover arguments used for further processing by other functions}
}
}
\description{
Train the SM2007 model
}
