% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TD.R
\name{TD}
\alias{TD}
\title{Train the TD model}
\usage{
TD(parameters, experience, mapping, debug = FALSE, debug_t = -1, ...)
}
\arguments{
\item{parameters}{A list containing the model parameters,
as returned by get_parameters().}

\item{experience}{A data.frame specifying trials as rows,
as returned by \code{make_experiment}}

\item{mapping}{A named list specifying trial and stimulus mapping,
as returned by \code{make_experiment}}
}
\value{
A list with raw results
}
\description{
Train the TD model
}
\note{
This model is in a highly experimental state. Use with caution.
}
