% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pavHEIDI.R
\name{train_pavHEIDI}
\alias{train_pavHEIDI}
\title{Trains a Pavlovian HeiDI}
\usage{
train_pavHEIDI(sals, cons, w, ts, trials, trialnames, targets = "US")
}
\arguments{
\item{sals}{A named vector with stimulus saliencies.}

\item{cons}{A named vector with stimulus constants.}

\item{w}{A named array of dimensions S,S; where S is the number of stimuli.}

\item{ts}{A vector of trial pointers for training, as a function of trials.}

\item{trials}{A list of length T, with character vectors specifying the stimuli involved in each trial. T is the number of unique trials in the experiment.}

\item{trialnames}{A character vector of length T with the names of the trials}

\item{targets}{A character vector specifying the target stimulus for each trial. Under development. For now, just the US.}
}
\value{
A list with
\itemize{
\item{ws - An array of dimensions P,S,S; where P is the number of trials used to train the model and S is the number of stimuli involved in the experiment.}
\item{rs,combvs, chainvs - Lists of length P with the r-values, combination vs, and chain vs.}
\item{ts, trials, trialnames -  Carryover for further processing. See Arguments.}
}
}
\description{
Trains a Pavlovian HeiDI
}
\note{
The array w contains the associations for all stimuli involved in the experiment. Entry i,j specifies the associative strength between stimulus i to stimulus j. Entry j,i specifies the opposite direction.
}
