% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_parser.R
\name{trial_parser}
\alias{trial_parser}
\title{Parses a string stipulating trials}
\usage{
trial_parser(trial_string)
}
\arguments{
\item{trial_string}{A string}
}
\value{
A list with
\itemize{
\item {trial_names: A character vector specifying the names of the trials contained in the string}
\item {trial_repeats: A numeric vector specifying the number of times each trial is given}
\item {trial_pre_functional: A list of character vectors specifying the unique functional stimuli on the expectation part of each trial}
\item {trial_post_functional: A list of character vectors specifying the unique functional stimuli on the correction part of each trial}
\item {trial_pre_nominal: A list of character vectors specifying the nominal stimuli on the expectation part of each trial}
\item {trial_post_nominal: A list of character vectors specifying the nominal stimuli on the expectation part of each trial}
\item {nomi_func_map: A data.frame specifying the mapping between functional and nominal stimuli}
\item {unique_nominal stimuli: A character vector with the unique nominal stimuli}
\item {unique_functional_stimuli: A character vector with the unique functional stimuli}
\item {is_test: A logical vector specifying whether each trial should not result in learning}
\item
}
}
\description{
Parses a string stipulating trials
}
\note{
Supports complex stimuli in parentheses (e.g., "5A>(US)")
Supports multiple nominal versions of stimuli (e.g., "3(A_a)>(US_a)/3(A_b)>(US_b)")
Supports probe trials followed by a hash (e.g., "3X#")
On trials that do not have the ">" character, it is assumed that all the specified stimuli are to be used in the expectation step. If so, the "post" entries in the list will contain NAs, as necessary.
On those trials, the "post" stimuli will be none (a character(0) vector).
}
\examples{
trial_parser("30A")
trial_parser("30XA>U/30(YA)>A")
trial_parser("1X#/2X>(US)#")
trial_parser("1X>(US_a)/1Y>(US_b)")
}
