% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialParser.R
\name{trialParser}
\alias{trialParser}
\title{Parses a string stipulating trials}
\usage{
trialParser(str)
}
\arguments{
\item{str}{A string}
}
\value{
A list with the number of trials, the trials, and the stimulus names
}
\description{
Parses a string stipulating trials
}
\examples{
trialParser("30XAU/30(YA)A")
}
