% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANCCR.R
\name{ANCCR}
\alias{ANCCR}
\title{Train the ANCCR model}
\usage{
ANCCR(parameters, experience, mapping, debug = FALSE, debug_t = -1, ...)
}
\arguments{
\item{parameters}{A list containing the model parameters,
as returned by get_parameters().}

\item{experience}{A data.frame specifying trials as rows,
as returned by \code{make_experiment}}

\item{mapping}{A named list specifying trial and stimulus mapping,
as returned by \code{make_experiment}}

\item{debug}{Logical specifying whether to print debug information.}

\item{debug_t}{Whether to invoke a \code{browser} at
the end of a timestep equal to debug_t.}

\item{...}{Additional named arguments}
}
\value{
A list with raw results
}
\description{
Train the ANCCR model
}
\note{
This model is in a highly experimental state. Use with caution.
}
