% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmr_app.R
\name{app}
\alias{app}
\title{Run the calmr shiny application}
\usage{
app(browser = getOption("shiny.launch.browser", interactive()))
}
\arguments{
\item{browser}{If TRUE, the app is launched in a browser.}
}
\description{
Run the calmr shiny application
}
