% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R
\name{show,CalmExperiment-method}
\alias{show,CalmExperiment-method}
\alias{design,CalmExperiment-method}
\alias{design}
\alias{arguments,CalmExperiment-method}
\alias{arguments}
\alias{c,CalmExperiment-method}
\alias{parameters<-}
\alias{parameters}
\alias{parameters,CalmExperiment-method}
\alias{parameters<-,CalmExperiment-method}
\alias{experience<-}
\alias{experience}
\alias{experience,CalmExperiment-method}
\alias{experience<-,CalmExperiment-method}
\alias{results,CalmExperiment-method}
\alias{results}
\alias{raw_results,CalmExperiment-method}
\alias{raw_results}
\alias{parsed_results,CalmExperiment-method}
\alias{parsed_results}
\alias{length,CalmExperiment-method}
\alias{parse}
\alias{parse,CalmExperiment-method}
\alias{aggregate,CalmExperiment-method}
\alias{plot,CalmExperiment-method}
\alias{plot}
\title{CalmExperiment methods}
\usage{
\S4method{show}{CalmExperiment}(object)

\S4method{design}{CalmExperiment}(x)

\S4method{arguments}{CalmExperiment}(x)

\S4method{c}{CalmExperiment}(x, ..., recursive = FALSE)

parameters(x) <- value

\S4method{parameters}{CalmExperiment}(x)

\S4method{parameters}{CalmExperiment}(x) <- value

experience(x) <- value

\S4method{experience}{CalmExperiment}(x)

\S4method{experience}{CalmExperiment}(x) <- value

\S4method{results}{CalmExperiment}(object)

\S4method{raw_results}{CalmExperiment}(object)

\S4method{parsed_results}{CalmExperiment}(object)

\S4method{length}{CalmExperiment}(x)

parse(object)

\S4method{parse}{CalmExperiment}(object)

\S4method{aggregate}{CalmExperiment}(x, ...)

\S4method{plot}{CalmExperiment}(x, type = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{CalmExperiment-class}}.}

\item{...}{Objects to concatenate}

\item{recursive}{Unused}

\item{value}{A list of experiences.}

\item{type}{character vector specifying the types of plots to create.
See ??supported_plots. Defaults to NULL.}
}
\value{
A ggplot object
}
\description{
Creates plots (or plot) with aggregated results in CalmExperiment
}
\note{
With type = NULL, all supported plots are returned.
}
\examples{
# Concatenate experiments
df <- get_design("blocking")
mods <- c("RW1972", "MAC1975")
exs <- lapply(mods, function(m) {
  make_experiment(df, parameters = get_parameters(df, model = m), model = m)
})
# joining separate experiments
c(exs[[1]], exs[[2]])
# joining a list of experiments
joint <- do.call(c, exs)
}
