% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R
\name{show,CalmExperiment-method}
\alias{show,CalmExperiment-method}
\alias{design,CalmExperiment-method}
\alias{design}
\alias{trials,CalmExperiment-method}
\alias{trials}
\alias{parameters<-}
\alias{parameters}
\alias{parameters,CalmExperiment-method}
\alias{parameters<-,CalmExperiment-method}
\alias{experiences,CalmExperiment-method}
\alias{experience}
\alias{results,CalmExperiment-method}
\alias{results}
\alias{raw_results,CalmExperiment-method}
\alias{raw_results}
\alias{parsed_results,CalmExperiment-method}
\alias{parsed_results}
\alias{length,CalmExperiment-method}
\alias{parse}
\alias{parse,CalmExperiment-method}
\alias{aggregate,CalmExperiment-method}
\alias{plot,CalmExperiment-method}
\alias{plot}
\title{CalmExperiment methods}
\usage{
\S4method{show}{CalmExperiment}(object)

\S4method{design}{CalmExperiment}(x)

\S4method{trials}{CalmExperiment}(object)

parameters(x) <- value

\S4method{parameters}{CalmExperiment}(x)

\S4method{parameters}{CalmExperiment}(x) <- value

\S4method{experiences}{CalmExperiment}(x)

\S4method{results}{CalmExperiment}(object)

\S4method{raw_results}{CalmExperiment}(object)

\S4method{parsed_results}{CalmExperiment}(object)

\S4method{length}{CalmExperiment}(x)

parse(object)

\S4method{parse}{CalmExperiment}(object)

\S4method{aggregate}{CalmExperiment}(x, ...)

\S4method{plot}{CalmExperiment}(x, type = NULL, ...)
}
\arguments{
\item{x}{An object of class \code{\link{CalmExperiment-class}}.}

\item{value}{A list of parameter lists.}

\item{type}{character vector specifying the types of plots to create.
See ??supported_plots. Defaults to NULL.}
}
\value{
A ggplot object
}
\description{
Creates plots (or plot) with aggregated results in CalmExperiment
}
\note{
With type = NULL, all supported plots are returned.
}
