% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R
\docType{class}
\name{CalmExperiment-class}
\alias{CalmExperiment-class}
\title{S4 class for calm experiments.}
\description{
S4 class for calm experiments.
}
\section{Slots}{

\describe{
\item{\code{arguments}:}{A list containing arguments to run models.}
\item{\code{design}:}{A CalmDesign object.}
\item{\code{results}:}{A CalmExperimentResult object.}
}
}

\seealso{
CalmExperiment-methods
}
