% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmr-package.R
\docType{package}
\name{calmr-package}
\alias{calmr}
\alias{calmr-package}
\title{calmr: Canonical Associative Learning Models and their Representations}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Implementations of canonical associative learning models, with tools to run experiment simulations, estimate model parameters, and compare model representations. Experiments and results are represented using S4 classes and methods.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/victor-navarro/calmr}
  \item \url{https://victornavarro.org/calmr/}
  \item Report bugs at \url{https://github.com/victor-navarro/calmr/issues}
}

}
\author{
\strong{Maintainer}: Victor Navarro \email{navarrov@cardiff.ac.uk} [copyright holder]

}
\keyword{internal}
