% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmr_function.R
\docType{package}
\name{CaLMR-package}
\alias{CaLMR}
\alias{CaLMR-package}
\title{Package imports}
\description{
R functions for conduting MR analysis for Causal analysis of Latent exposures using Mendelian Randomization (CaLMR).
}
\keyword{internal}
