% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_plots.R
\name{calmr_model_plot}
\alias{calmr_model_plot}
\title{Create a plot with calm data}
\usage{
calmr_model_plot(dat, type)
}
\arguments{
\item{dat}{A tbl with data to use in the plot}

\item{type}{A character specifying the type of plot}
}
\value{
A ggplot object
}
\description{
Create a plot with calm data
}
\note{
You should probably be getting plots via
the \code{plot} method for CalmrExperiments.
}
