% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaLMR(Multi).R
\name{calmr_multi}
\alias{calmr_multi}
\title{Bayesian Mendelian Randomization Analysis for Latent Exposures Leveraging GWAS Summary Statistics for Traits Co-Regulated by the Exposures ##
This function conducts the Multivariable (multi-exposure) Causal analysis of Latent exposures using Mendelian Randomization ( CaLMR(Multi) )
The function uses summary-level association statistics from genome-wide association study}
\usage{
calmr_multi(
  sumtable,
  Corr.mat,
  grp,
  L,
  K,
  traitvec,
  outcome,
  sign,
  T = 3000,
  burnin = 1500
)
}
\arguments{
\item{sumtable}{a M*(K+1) data frame containing the GWAS summary data for K observable traits and the outcome.}

\item{Corr.mat}{a (K+1)*(K+1) estimated correlation matrix of the GWAS summary statistics.
The 1st-Kth variables are related to the K observable traits, and the last variable corresponds to the outcome.
*The order of the observable traits should match with the order in the 'traitvec' vector.}

\item{grp}{a list of length L with the sub-list grp[[l]] containing the names of the observable traits associated with l-th latent exposure.}

\item{L}{number of latent exposures}

\item{K}{number of observable traits}

\item{traitvec}{a vector containing the names of the observable traits. This should match with the column names in sumtable.}

\item{outcome}{the name of the outcome Y, and this should match with the column name in sumtable.}

\item{sign}{a list of length L. Each sub-list sign[[l]] should be a vector of length K containing the pre-known signs of theta_kl.
[1 means positive, -1 means negative, and 0 means this trait is not associated with l-th latent exposure or lack of pre-known signs]}

\item{T}{total number of iterations for the Gibbs sampler, with default T=3000}

\item{burnin}{length of burn-in period, with default burnin=1500.}
}
\value{
a list of the testing results based on CaLMR
       - bayes.rej: a vector of length L showing the existences of causal effects for each latent exposure
       - bayes.sign: a vector of length L showing the the directions of the significant causal effects. 0 means no causal effects.
       - cor.X: estimated correlation matrix for the latent exposures
       - par: posterior samples
       - ci: 95% credible intervals for model parameters after dropping the burnin period.
}
\description{
Bayesian Mendelian Randomization Analysis for Latent Exposures Leveraging GWAS Summary Statistics for Traits Co-Regulated by the Exposures ##
This function conducts the Multivariable (multi-exposure) Causal analysis of Latent exposures using Mendelian Randomization ( CaLMR(Multi) )
The function uses summary-level association statistics from genome-wide association study
}
