% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaLMR(Uni).R
\name{calmr_uni}
\alias{calmr_uni}
\title{Univariable (single-exposure) Causal analysis of Latent exposures using Mendelian Randomization CaLMR(Uni)}
\usage{
calmr_uni(
  sumtable,
  Corr.mat,
  K = K,
  traitvec,
  outcome,
  sign,
  T = 3000,
  burnin = 1500
)
}
\arguments{
\item{sumtable}{a M*(K+1) data frame containing the GWAS summary data for K observable traits and the outcome.}

\item{Corr.mat}{a (K+1)*(K+1) estimated correlation matrix of the GWAS summary statistics.
The 1st-Kth variables are related to the K observable traits, and the last variable corresponds to the outcome.
*The order of the observable traits should match with the order in the 'traitvec' vector.}

\item{K}{number of observable traits}

\item{traitvec}{a vector containing the names of the observable traits. This should match with the column names in sumtable.}

\item{outcome}{the name of the outcome Y, and this should match with the column name in sumtable.}

\item{sign}{a vector of length K containing the pre-known signs of theta_k.
[1 means positive, -1 means negative, and 0 means this trait is not associated with l-th latent exposure or lack of pre-known signs]}

\item{T}{total number of iterations for the Gibbs sampler, with default T=3000}

\item{burnin}{length of burn-in period, with default burnin=1500.}
}
\value{
a list containing the CaLMR result
 \describe{
  \item{bayes.rej}{the existences of causal effect between the latent exposure and the outcome}
  \item{bayes.sign}{the direction of the significant causal effect. 0 means no causal effect}
  \item{par}{posterior samples generated by the Gibbs sampling using an MCMC algorithm}
  \item{ci}{95\% credible intervals for model parameters after dropping the burnin period}
 }
}
\description{
Causal analysis of Latent exposures using Mendelian Randomization (CaLMR) is an MR method that tests the causal relationships between
the outcome and the latent exposure using GWAS summary-level association statistics.
This function conducts CaLMR(Uni) test, assuming there is one latent exposure.
It is built under a two-sample MR framework and conducts Bayesian modeling using conjugate priors and Regression with Summary Statistics (RSS) Likelihood.
}
