% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calmr_verbosity.R
\name{calmr_verbosity}
\alias{calmr_verbosity}
\title{Set verbosity options for calmr}
\usage{
calmr_verbosity(verbose)
}
\arguments{
\item{verbose}{A logical}
}
\value{
The list of progressr handlers (invisibly).
}
\description{
Whether to show verbosity messages and progress bars
}
\note{
Progress bars are handled by the progressr package.
This is just a convenience function.

See package 'progressr' for further details.
}
