% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design.R
\docType{class}
\name{CalmrDesign-class}
\alias{CalmrDesign-class}
\title{S4 class for calmr designs}
\description{
S4 class for calmr designs
}
\section{Slots}{

\describe{
\item{\code{design}:}{A list containing design information.}
\item{\code{mapping}:}{A list containing the object mapping.}
\item{\code{raw_design}:}{The original data.frame.}
}
}

