% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design.R, R/anccr_helpers.R
\name{augment,CalmrDesign-method}
\alias{augment,CalmrDesign-method}
\alias{augment}
\alias{.anccrize_design}
\alias{show,CalmrDesign-method}
\alias{mapping}
\alias{mapping,CalmrDesign-method}
\alias{trials}
\alias{trials,CalmrDesign-method}
\title{Augment CalmrDesign}
\usage{
\S4method{augment}{CalmrDesign}(object, model, ...)

.anccrize_design(object)

\S4method{show}{CalmrDesign}(object)

mapping(object)

\S4method{mapping}{CalmrDesign}(object)

trials(object)

\S4method{trials}{CalmrDesign}(object)
}
\arguments{
\item{object}{A CalmrDesign, as returned by parse_design}

\item{model}{A modelname string. One of supported_models()}

\item{...}{Additional parameters depending on the model.}
}
\description{
Methods mapping, and trials, are extractor functions.
}
