% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_design.R
\name{CalmrDesign-methods}
\alias{CalmrDesign-methods}
\alias{show,CalmrDesign-method}
\alias{mapping,CalmrDesign-method}
\alias{mapping}
\alias{trials,CalmrDesign-method}
\title{CalmrDesign methods}
\usage{
\S4method{show}{CalmrDesign}(object)

\S4method{mapping}{CalmrDesign}(object)

\S4method{trials}{CalmrDesign}(object)
}
\arguments{
\item{object}{A \code{CalmrDesign} object}
}
\value{
\code{show()} returns NULL (invisibly).

\code{mapping()} returns a list with trial mappings.

\code{trials()} returns NULL (invisibly).
}
\description{
S4 methods for \code{CalmrDesign} class.
}
