% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_result.R
\name{show,CalmResult-method}
\alias{show,CalmResult-method}
\title{Methods for CalmResult}
\usage{
\S4method{show}{CalmResult}(object)
}
\arguments{
\item{object}{A CalmResult object}
}
\description{
Methods for CalmResult
}
