% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R
\name{show,CalmrExperiment-method}
\alias{show,CalmrExperiment-method}
\alias{design,CalmrExperiment-method}
\alias{design}
\alias{c,CalmrExperiment-method}
\alias{parameters,CalmrExperiment-method}
\alias{parameters}
\alias{parameters<-,CalmrExperiment-method}
\alias{results,CalmrExperiment-method}
\alias{results}
\alias{length,CalmrExperiment-method}
\alias{parse,CalmrExperiment-method}
\alias{aggregate,CalmrExperiment-method}
\title{CalmrExperiment methods}
\usage{
\S4method{show}{CalmrExperiment}(object)

\S4method{design}{CalmrExperiment}(x)

\S4method{c}{CalmrExperiment}(x, ..., recursive = FALSE)

\S4method{parameters}{CalmrExperiment}(x)

\S4method{parameters}{CalmrExperiment}(x) <- value

\S4method{results}{CalmrExperiment}(object)

\S4method{length}{CalmrExperiment}(x)

\S4method{parse}{CalmrExperiment}(object)

\S4method{aggregate}{CalmrExperiment}(x, ...)
}
\description{
CalmrExperiment methods
}
\examples{
# Concatenate designs
df <- get_design("basic_blocking")
pars1 <- get_parameters(df, model = "RW1972")
pars2 <- get_parameters(df, model = "MAC1975")
ex1 <- make_experiment(df, model = "RW1972")
ex2 <- make_experiment(df, model = "MAC1975")
c(ex1, ex2)

# Concatenate results
lapply(c(ex1, ex2), run_experiment)
}
