% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CalmrExperiment}
\alias{CalmrExperiment}
\title{CalmrExperiment}
\description{
S4 classes for calmr experiments.
}
\section{Slots}{

\describe{
\item{\code{.Data}:}{Inherited from tbl class}
\item{\code{raw_results}:}{List. The raw results.}
\item{\code{is_parsed}:}{Logical. Whether the model results have been parsed}
}
}

