% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R
\docType{class}
\name{CalmrExperiment-class}
\alias{CalmrExperiment-class}
\title{S4 class for calm experiments.}
\description{
S4 class for calm experiments.
}
\section{Slots}{

\describe{
\item{\code{arguments}:}{A tbl containing arguments to run models.}
\item{\code{design}:}{A CalmrDesign object.}
\item{\code{results}:}{A CalmrExperimentResult object.}
}
}

\seealso{
CalmrExperiment-methods
}
