% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CalmrModel-class}
\alias{CalmrModel-class}
\title{S4 class for Calmr Models}
\description{
S4 class for Calmr Models
}
\section{Slots}{

\describe{
\item{\code{model}:}{Character. The model name.}
\item{\code{parameters}:}{List. The model parameters.}
\item{\code{model_results }:}{List. The model results (layers)}
\item{\code{experience}:}{Data.frame. The data.frame used to train the model.}
\item{\code{mapping}:}{List. Contains mapping information (e.g., between functional and nominal stimuli)}
\item{\code{is_parsed}:}{Logical. Whether the model results have been parsed.}
}
}

