% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\docType{class}
\name{CalmrModel-class}
\alias{CalmrModel-class}
\title{S4 class for calmr Models}
\description{
S4 class for calmr Models
}
\section{Slots}{

\describe{
\item{model}{A model name string}
\item{parameters}{A list with the model with model parameters}
}
}

