% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\docType{class}
\name{CalmrModel-class}
\alias{CalmrModel-class}
\title{S4 class for Calmr Models}
\description{
S4 class for Calmr Models
}
\note{
Currently unused.
}
\section{Slots}{

\describe{
\item{model}{A model name string}
\item{model_fn}{The model function}
\item{state}{A list with the model internal state}
}
}

