% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_model.R
\name{CalmrModel-methods}
\alias{CalmrModel-methods}
\alias{run,CalmrModel-method}
\alias{run}
\alias{parameters,CalmrModel-method}
\alias{parameters<-,CalmrModel-method}
\alias{results,CalmrModel-method}
\alias{show,CalmrModel-method}
\alias{parse,CalmrModel-method}
\title{CalmrModel methods}
\usage{
\S4method{run}{CalmrModel}(object, experience, mapping, ...)

\S4method{parameters}{CalmrModel}(x)

\S4method{parameters}{CalmrModel}(x) <- value

\S4method{results}{CalmrModel}(object)

\S4method{show}{CalmrModel}(object)

\S4method{parse}{CalmrModel}(object, outputs = object@outputs)
}
\arguments{
\item{object}{A \link{CalmrModel} object.}

\item{experience}{A data.frame specifying trials as rows,
as returned by \code{make_experiment()}.}

\item{mapping}{A named list specifying trial and stimulus mapping,
as returned by \code{make_experiment()}.}

\item{...}{Additional named arguments.}
}
\value{
\code{run()} returns the updated \link{CalmrModel} object.

\code{parameters()} returns the parameters of the \link{CalmrModel} object.

\verb{parameters()<-} sets the parameters of a \link{CalmrModel} object.

results() returns the last results of the \link{CalmrModel} object.

show() returns NULL (invisibly).

parse() returns a list with parsed results.
}
\description{
S4 methods for \code{CalmrModel} class.

Run a \link{CalmrModel} object with the given parameters,
experience, and mapping.
}
\note{
The \code{run} method changes some internal
states of the model (if appropriate) and
populates the \code{.last_results} slot with the results of the run.
}
