% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_result.R
\name{show,CalmrResult-method}
\alias{show,CalmrResult-method}
\alias{c,CalmrExperimentResult-method}
\title{Methods for CalmrResult}
\usage{
\S4method{show}{CalmrResult}(object)

\S4method{c}{CalmrExperimentResult}(x, ..., recursive = FALSE)
}
\description{
Methods for CalmrResult
}
