% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_result.R
\name{CalmrResult-methods}
\alias{CalmrResult-methods}
\alias{show,CalmrResult-method}
\title{CalmrResult methods}
\usage{
\S4method{show}{CalmrResult}(object)
}
\arguments{
\item{object}{A \code{CalmrResults} object.}
}
\value{
\itemize{
\item \code{show()} returns NULL (invisibly).
}
}
\description{
S4 methods for \code{CalmrResults} class.
}
