% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rsa.R
\docType{class}
\name{CalmrRSA-class}
\alias{CalmrRSA-class}
\title{S4 class for calmr representational similarity analysis}
\description{
S4 class for calmr representational similarity analysis
}
\section{Slots}{

\describe{
\item{\code{corr_mat}:}{An array containing the correlation matrix}
\item{\code{distances}:}{A list of pairwise distance matrices}
\item{\code{args}:}{A list of the arguments used to create the object.}
\item{\code{test_data}:}{A list with permutation data,
only populated after testing the object.}
}
}

