% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rsa.R
\name{CalmrRSA-methods}
\alias{CalmrRSA-methods}
\alias{show,CalmrRSA-method}
\alias{test,CalmrRSA-method}
\alias{test}
\alias{plot,CalmrRSA-method}
\title{CalmrRSA methods}
\usage{
\S4method{show}{CalmrRSA}(object)

\S4method{test}{CalmrRSA}(object, n_samples = 1000, p = 0.95)

\S4method{plot}{CalmrRSA}(x)
}
\arguments{
\item{object, x}{A \code{CalmrRSA} object.}

\item{n_samples}{The number of samples for the permutation test
(default = 1e3)}

\item{p}{The critical threshold level for the permutation test
(default = 0.95)}
}
\value{
\itemize{
\item \code{show()} returns NULL (invisibly).
\item \code{test()} returns the \code{CalmrRSA} object with permutation test data.
\item \code{plot()} returns a list of 'ggplot' plot objects.
}
}
\description{
S4 methods for \code{CalmrRSA} class.
}
