% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_design.R
\name{get_design}
\alias{get_design}
\title{Get basic designs}
\usage{
get_design(design_name = NULL)
}
\arguments{
\item{design_name}{A string specifying a design name (default = NULL)}
}
\value{
If design_name is not NULL, a data.frame containing the design.
Otherwise, a list containing all available designs.
}
\description{
Get basic designs
}
\examples{
names(get_design())
get_design("blocking")
}
\seealso{
\code{\link[=parse_design]{parse_design()}}
}
