% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_heidi_opts.R
\name{get_heidi_opts}
\alias{get_heidi_opts}
\title{Get options for heidi model}
\usage{
get_heidi_opts(iterations = 1, miniblocks = TRUE)
}
\arguments{
\item{iterations}{An integer denoting the number of iterations to run. Default = 1.}

\item{miniblocks}{A logical denoting whether to create trial miniblocks when possible. Default = TRUE.}
}
\description{
Get options for heidi model
}
\examples{
get_heidi_opts(iterations = 10)
}
