% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model.R
\name{get_model}
\alias{get_model}
\title{Return a model function}
\usage{
get_model(model_name)
}
\arguments{
\item{model_name}{A string specifying the model to be returned. One of \verb{c("HDI2020", "HD2022", "RW1972)}.}
}
\description{
Given a model string, returns a model function that can be used elsewhere in the package.
}
\note{
The model_name parameter matches published models as follows:
\itemize {
\item {HDI2020: Honey, Dwyer, and Illiescu (2020). Psych Review.}
\item {HD2022: Honey and Dwyer (2022). Psych Review.}
}
}
