% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_output}
\alias{get_output}
\title{Get output from CalmrExperiment}
\usage{
get_output(object, type = NULL)
}
\arguments{
\item{object}{An object of class \code{\link{CalmrExperiment-class}}.}

\item{type}{The type of output}
}
\value{
A data.frame with the model output
}
\description{
Returns a tibble containing parsed outputs of CalmrExperiment
}
