% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Get model parameters}
\usage{
get_parameters(design, model = NULL)
}
\arguments{
\item{design}{A \code{data.frame} containing the experimental design.}

\item{model}{A string specifying a model. One in \code{\link[=supported_models]{supported_models()}}.}
}
\value{
A list with model parameters depending on model
}
\description{
Get model parameters
}
\examples{
block <- get_design("blocking")
get_parameters(block, model = "SM2007")
}
