% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timings.R
\name{get_timings}
\alias{get_timings}
\title{Get timing design parameters}
\usage{
get_timings(design, model)
}
\arguments{
\item{design}{A \code{data.frame} containing the experimental design.}

\item{model}{One of \code{\link[=supported_timed_models]{supported_timed_models()}}.}
}
\value{
A list of timing design parameters.
}
\description{
Get timing design parameters
}
\examples{
block <- get_design("blocking")
get_timings(block, model = "TD")
}
