% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_graphs.R, R/class_experiment.R
\name{calmr_model_graph}
\alias{calmr_model_graph}
\alias{graph,CalmrExperiment-method}
\title{Create a graph with calm data}
\usage{
calmr_model_graph(
  x,
  loops = TRUE,
  limits = max(abs(x$value)) * c(-1, 1),
  colour_key = FALSE,
  t = max(x$trial),
  options = get_graph_opts(),
  ...
)

\S4method{graph}{CalmrExperiment}(x, ...)
}
\arguments{
\item{x}{A CalmrExperiment}

\item{loops}{Logical. Whether to draw arrows back and forth}

\item{limits}{Numerical. Limits for color scale. Defaults to NULL,
in which case, limits are set to be max(abs(x$value))*c(-1,1)}

\item{colour_key}{Logical. Whether to show the color key}

\item{t}{The trial from which weights are obtained
(defaults to the maximum trial in the design)}

\item{options}{A list with graph options, as returned by \code{get_graph_opts}}

\item{...}{Additional parameters passed to \code{graph_calmr_model}}

\item{trial}{Numerical. The trial to graph.}
}
\value{
A ggplot object
}
\description{
Create a graph with calm data

Graph CalmrExperiment
}
\note{
You should probably be getting graphs via
the \code{graph} method for CalmrExperiments.
}
