% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HD2022.R
\name{HD2022}
\alias{HD2022}
\title{Train the HD2022 model}
\usage{
HD2022(v = NULL, parameters, experience, mapping, ...)
}
\arguments{
\item{v}{(optional) A named matrix of dimensions S,S;
where S is the number of stimuli.}

\item{parameters}{A list containing the model parameters,
as returned by get_parameters().}

\item{experience}{A data.frame specifying trials as rows,
as returned by \code{make_experiment}}

\item{mapping}{A named list specifying trial and stimulus mapping,
as returned by \code{make_experiment}}

\item{...}{Additional named arguments}
}
\value{
A list with raw results
}
\description{
Train the HD2022 model
}
