% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heidi_app.R
\name{heidi_app}
\alias{heidi_app}
\title{Run the heidi GUI}
\usage{
heidi_app(browser = FALSE)
}
\arguments{
\item{browser}{A logical specifying whether the app should be launched into a browser. Default = FALSE.}
}
\description{
Starts a shiny application
}
