% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heidi_helpers.R
\name{heidi_helpers}
\alias{heidi_helpers}
\alias{genSSWeights}
\alias{parseWs}
\alias{parseVs}
\alias{parseRs}
\title{An assortment of functions to help heidi}
\usage{
genSSWeights(stims, default_val = 0)

parseWs(mod)

parseVs(mod)

parseRs(mod)
}
\arguments{
\item{stims}{A character vector with stimuli.}

\item{default_val}{Default alpha value.}

\item{mod}{A model list, as returned by trainPavHEIDI.}
}
\description{
genSSWeights generates a named array with model weights. parseWs, parseVs, and parseRs parse the raw outputs of trainPavHEIDI into a readable format.
}
