% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heidi_helpers.R
\name{heidi_helpers}
\alias{heidi_helpers}
\alias{gen_ss_weights}
\alias{parse_ws}
\alias{parse_vs}
\alias{parse_rs}
\alias{parse_heidi_results}
\title{An assortment of functions to help heidi}
\usage{
gen_ss_weights(stims, default_val = 0)

parse_ws(mod)

parse_vs(mod)

parse_rs(mod)

parse_heidi_results(raw_results)
}
\arguments{
\item{stims}{A character vector with stimuli.}

\item{default_val}{Default alpha value.}

\item{mod}{A model list, as returned by trainPavHEIDI.}

\item{raw_results}{A tibble with model information, as return by run_heidi}
}
\description{
gen_ss_weights generates a named array with model weights. parse_ws, parse_vs, parse_rs and parse_heidi_results parse the raw outputs of trainPavHEIDI into a readable format.
}
