% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heidi_helpers.R
\name{heidi_helpers}
\alias{heidi_helpers}
\alias{get_params}
\alias{gen_ss_weights}
\alias{parse_ws}
\alias{parse_vs}
\alias{parse_rs}
\alias{parse_as}
\alias{parse_heidi_results}
\alias{filter_heidi_results}
\title{An assortment of functions to help heidi}
\usage{
get_params(design, default_par = 0.2)

gen_ss_weights(stims, default_val = 0)

parse_ws(mod)

parse_vs(mod)

parse_rs(mod)

parse_as(mod)

parse_heidi_results(raw_results)

filter_heidi_results(parsed_results, filters)
}
\arguments{
\item{design}{An experimental design. Either a data.frame or a tibble returned by parse_design}

\item{default_par}{A float between 0 and 1}

\item{stims}{A character vector with stimuli}

\item{default_val}{Default alpha value}

\item{mod}{A model list, as returned by train_pav_heidi}

\item{raw_results}{A tibble with model information, as returned by quick_heidi}

\item{parsed_results}{A list with parsed results, as returned by parse_heidi_results}

\item{filters}{A named list containing "phase" and "trial_type" character vectors, for filtering data}
}
\description{
get_params generates a data.frame with stimulus salience parameters.
gen_ss_weights generates a named array with model weights.
parse_ws, parse_vs, parse_rs and parse_heidi_results parse the raw outputs of train_pav_heidi into a readable format.
filter_heidi_results is a convenience function to filter specific phase and trial_type data.
}
