% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heidi_plots.R
\name{heidi_plots}
\alias{heidi_plots}
\alias{make_plots}
\alias{plot_ws}
\alias{plot_vs}
\alias{plot_rs}
\alias{plot_as}
\alias{patch_plots}
\alias{get_plot_opts}
\alias{plot_common_scale}
\alias{graph_weights}
\alias{get_graph_opts}
\alias{make_graphs}
\alias{patch_graphs}
\title{Plotting functions for heidi models}
\usage{
make_plots(dat)

plot_ws(vals)

plot_vs(vals, bars = F)

plot_rs(vals, simple = F)

plot_as(vals)

patch_plots(plots, selection, options = get_plot_opts())

get_plot_opts(common_scale = TRUE)

plot_common_scale(plots)

graph_weights(
  ws,
  limits = max(abs(range(ws$value))) * c(-1, 1),
  t = max(ws$trial),
  opts = get_graph_opts()
)

get_graph_opts(
  arrow.gap = 0.16,
  arrow.curvature = 0.2,
  arrow.pt = 20,
  edge.size = 3,
  node.size = 40,
  node.stroke = 3,
  node.text.size = 15
)

make_graphs(
  mod,
  limits = max(abs(range(mod$ws$value))) * c(-1, 1),
  trial = max(mod$ws$trial)
)

patch_graphs(graphs)
}
\arguments{
\item{dat}{A list containing data.frames with parsed weights, r values, and v values, as returned by parse_heidi_results.}

\item{vals}{A data.frame containing parsed values}

\item{bars}{A logical stipulating whether to summarize and use stacked bars, instead of points and lines.}

\item{simple}{A logical stipulating whether to simplify the plot by collapsing across sources.}

\item{plots}{A named list with plots}

\item{selection}{A character vector with the selected plots}

\item{options}{A list with options}

\item{common_scale}{A logical. Whether to plot the data in a common y-scale.}

\item{ws}{A data.frame containing parsed weights, as returned by parse_heidi_results}

\item{limits}{A vector of length 2 specifying the range of weights. Defaults to the negative and positive maximum of absolute weights.}

\item{t}{An integer specifying a trial}

\item{opts}{A list of options for graphing weights. See ?get_graph_opts.}

\item{arrow.gap}{Distance between nodes and arrows}

\item{arrow.curvature}{Curvature of the arrows}

\item{arrow.pt}{Size of the arrows (see grid::arrow)}

\item{edge.size}{Thickness of the arrows}

\item{node.size}{Size of the nodes}

\item{node.stroke}{Thickness of the node lines}

\item{node.text.size}{Thickness of the text within the node}

\item{mod}{Parsed model results, as returned by parse_heidi_results.}

\item{trial}{An integer denoting the trial of the weights to be graphed. Defaults to the last trial in the data.}

\item{graphs}{A list of graphs, as returned by make_graphs}
}
\value{
A list with ggplot objects
}
\description{
Plotting functions for heidi models
}
\note{
\itemize{
\item{Plotting options are obtained via get_plot_opts(). For now, only plotting in a common y-axis is supported.
}}
}
\seealso{
parse_heidi_results
}
