% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heidi_plots.R
\name{heidi_plots}
\alias{heidi_plots}
\alias{make_plots}
\alias{plot_weights}
\alias{plot_vs}
\alias{plot_rs}
\alias{patch_plots}
\title{Plotting functions for heidi models}
\usage{
make_plots(dat)

plot_weights(vals)

plot_vs(vals)

plot_rs(vals)

patch_plots(plots, selection, options)
}
\arguments{
\item{dat}{A list containing data.frames with parsed weights, r values, and v values, as returned by parse_heidi_results.}

\item{vals}{A data.frame containing parsed values}

\item{plots}{A named list with plots}

\item{selection}{A character vector with the selected plots}

\item{options}{A character vector with all available plots}
}
\value{
A list with ggplot objects
}
\description{
Plotting functions for heidi models
}
\seealso{
parse_heidi_results
}
