% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heidi_plots.R
\name{heidi_plots}
\alias{heidi_plots}
\alias{make_plots}
\alias{plot_vs}
\alias{plot_acts}
\alias{plot_rs}
\alias{plot_as}
\alias{get_plot_opts}
\alias{plot_common_scale}
\alias{graph_weights}
\alias{get_graph_opts}
\alias{make_graphs}
\alias{patch_graphs}
\alias{patch_plots}
\title{Plotting functions for heidi models}
\usage{
make_plots(dat)

plot_vs(vals)

plot_acts(vals, bars = F)

plot_rs(vals, simple = F)

plot_as(vals)

get_plot_opts(common_scale = TRUE)

plot_common_scale(plots)

graph_weights(
  vs,
  limits = max(abs(range(vs$value))) * c(-1, 1),
  t = max(vs$trial),
  opts = get_graph_opts()
)

get_graph_opts(graph_size = "small")

make_graphs(
  mod,
  limits = max(abs(range(mod$vs$value))) * c(-1, 1),
  trial = max(mod$vs$trial),
  opts = get_graph_opts()
)

patch_graphs(graphs, selection = names(graphs))

patch_plots(plots, selection, options = get_plot_opts())
}
\arguments{
\item{dat}{A list containing data.frames with parsed weights, r values, and activation values, as returned by parse_heidi_results.}

\item{vals}{A data.frame containing parsed values}

\item{bars}{A logical stipulating whether to summarize and use stacked bars, instead of points and lines.}

\item{simple}{A logical stipulating whether to simplify the plot by collapsing across sources.}

\item{common_scale}{A logical. Whether to plot the data in a common y-scale.}

\item{plots}{A named list with plots}

\item{vs}{A data.frame containing parsed weights, as returned by parse_heidi_results}

\item{limits}{A vector of length 2 specifying the range of weights. Defaults to the negative and positive maximum of absolute weights.}

\item{t}{An integer specifying a trial}

\item{opts}{A list of options for graphing weights. See ?get_graph_opts.}

\item{graph_size}{A string specifying the desired graph size, from c("large", "small"). Default is "large"}

\item{mod}{Parsed model results, as returned by parse_heidi_results.}

\item{trial}{An integer denoting the trial of the weights to be graphed. Defaults to the last trial in the data.}

\item{graphs}{A list of graphs, as returned by make_graphs}

\item{selection}{A character vector with the selected plots}

\item{options}{A list with options}
}
\value{
A list with ggplot objects
}
\description{
Plotting functions for heidi models
}
\note{
\itemize{
\item{Plotting options are obtained via get_plot_opts(). For now, only plotting in a common y-axis is supported.
}}
}
\seealso{
parse_heidi_results
}
