% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heidi_plots.R
\name{heidi_plots}
\alias{heidi_plots}
\alias{make_plots}
\alias{plot_ws}
\alias{plot_vs}
\alias{plot_rs}
\alias{plot_as}
\alias{patch_plots}
\alias{get_plot_opts}
\alias{plot_common_scale}
\alias{graph_weights}
\alias{get_graph_opts}
\title{Plotting functions for heidi models}
\usage{
make_plots(dat)

plot_ws(vals)

plot_vs(vals, bars = F)

plot_rs(vals)

plot_as(vals)

patch_plots(plots, selection, options = get_plot_opts())

get_plot_opts(common_scale = TRUE)

plot_common_scale(plots)

graph_weights(mod, t = max(mod$ws$trial), opts = get_graph_opts())

get_graph_opts(
  arrow.gap = 0.16,
  arrow.curvature = 0.2,
  arrow.pt = 20,
  edge.size = 3,
  node.size = 40,
  node.stroke = 3,
  node.text.size = 15
)
}
\arguments{
\item{dat}{A list containing data.frames with parsed weights, r values, and v values, as returned by parse_heidi_results.}

\item{vals}{A data.frame containing parsed values}

\item{bars}{A logical stipulating whether to summarize and use stacked bars, instead of points and lines.}

\item{plots}{A named list with plots}

\item{selection}{A character vector with the selected plots}

\item{options}{A list with options}

\item{common_scale}{A logical. Whether to plot the data in a common y-scale.}

\item{ws}{A data.frame containing parsed weights, as returned by parse_heidi_results}

\item{trial}{An integer denoting the trial of the weights to be graphed. Defaults to the last trial in the data.}
}
\value{
A list with ggplot objects
}
\description{
Plotting functions for heidi models
}
\note{
\itemize{
\item{Plotting options are obtained via get_plot_opts(). For now, only plotting in a common y-axis is supported.
}}
}
\seealso{
parse_heidi_results
}
