% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeHeidiArgs.R
\name{make_heidi_args}
\alias{make_heidi_args}
\title{Make a tibble to fit a heidi model}
\usage{
make_heidi_args(design, pars, opts)
}
\arguments{
\item{design}{A design list, as returned by parse_design}

\item{pars}{A parameter data.frame as returned by get_params}

\item{opts}{A list with options}
}
\value{
A tibble with the arguments required to run the model. Each row represents a group in the experimental design.
}
\description{
Make a tibble to fit a heidi model
}
\examples{
df <- data.frame(Group = c('Group 1', 'Group 2'), P1 = c('10AB(US)', '10A(US)'), R1 = c(TRUE, TRUE))
des <- parse_design(df)
ps <- get_params(des, 0.2)
make_heidi_args(design = des, pars = ps, opts = list(iterations = 1))

}
\seealso{
parse_design, get_params
}
