% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_model_args.R
\name{make_model_args}
\alias{make_model_args}
\title{Make a tibble to fit a calmr model}
\usage{
make_model_args(design, pars = NULL, model = NULL, opts = get_exp_opts())
}
\arguments{
\item{design}{A design data.frame}

\item{pars}{A data.frame containing parameters as returned by \code{get_model_params}}

\item{model}{A string specifying the model name. One of \code{supported_models()}}

\item{opts}{A list with options as returned by \code{get_exp_opts}}
}
\value{
A tibble with the arguments required to run the model. Each row represents a group in the experimental design.
}
\description{
Make a tibble to fit a calmr model
}
\examples{
des <- data.frame(Group = "G1", P1 = "10A>(US)", R1 = TRUE)
ps <- get_model_params(des, model = "HD2022")
op <- get_exp_opts(iterations = 1)
make_model_args(design = des, pars = ps, model = "HD2022", opts = op)

}
\seealso{
\code{\link{parse_design}}, \code{\link{get_model_params}}, \code{\link{get_exp_opts}}
}
