% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_graphs.R
\name{patch_graphs}
\alias{patch_graphs}
\title{Patch Calmr graphs}
\usage{
patch_graphs(graphs, selection = names(graphs))
}
\arguments{
\item{graphs}{A list of named graphs, as returned by \code{\link[=graph]{graph()}} or
\code{\link[=calmr_model_graph]{calmr_model_graph()}}}

\item{selection}{A character or numeric vector
determining the plots to patch.}
}
\value{
A 'patchwork' object
}
\description{
Convenience function to patch graphs with 'patchwork'
}
