% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pati}
\alias{pati}
\title{Rat responses from Patittucci et al. 2016}
\format{
A data.frame with the following variables:
\describe{
\item{subject}{subject identifier}
\item{block}{the 2-session block of training (1 to 8)}
\item{lever}{lever presented on the trial: L = left; R = right}
\item{us}{the stimulus that followed the lever: P = pellet; S = sucrose}
\item{response}{the response: lp = lever press; np = nose poke}
\item{rpert}{responses per trial}
...
}
}
\source{
{Patittucci et al. (2016). JEP:ALC}
}
\usage{
pati
}
\description{
A dataset containing rat nose pokes and lever
presses when levers were associated with different appetitive stimuli.
}
\keyword{datasets}
