% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKH1982.R
\name{PKH1982}
\alias{PKH1982}
\title{Train the PKH1982 model}
\usage{
PKH1982(ev = NULL, iv = NULL, parameters, experience, mapping, ...)
}
\arguments{
\item{ev}{(optional) A named matrix of dimensions S,S;
where S is the number of stimuli.}

\item{iv}{(optional) A named matrix of dimensions S,S;
where S is the number of stimuli.}

\item{parameters}{A list containing the model parameters,
as returned by parameter_info().}

\item{experience}{A data.frame specifying trials as rows,
as returned by \code{make_experiment}}

\item{mapping}{A named list specifying trial and stimulus mapping,
as returned by \code{make_experiment}}
}
\value{
A list with raw results
}
\description{
Train the PKH1982 model
}
