% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R
\name{plot}
\alias{plot}
\title{Plot CalmrExperiment}
\usage{
plot(x, y, ...)
}
\arguments{
\item{x}{An object of class \code{\link{CalmrExperiment}}.}

\item{type}{character vector specifying the types of plots to create.
See ??supported_plots. Defaults to NULL.}
}
\value{
A ggplot object
}
\description{
Creates plots (or plot) with aggregated results in CalmrExperiment
}
\note{
With type = NULL, all supported plots are returned.
}
