% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAND.R
\name{RAND}
\alias{RAND}
\title{Train the RAND model}
\usage{
RAND(alphas, V = NULL, experience, mapping)
}
\arguments{
\item{alphas}{A named vector with stimulus saliences.}

\item{V}{(optional) A named matrix of dimensions S,S; where S is the number of stimuli.}

\item{experience}{A data.frame specifying trials as rows, as returned by \code{make_model_args}}

\item{mapping}{A named list specifying trial and stimulus mapping, as returned by \code{make_model_args}}

\item{lambdas}{A named vector with stimulus asymptotes.}
}
\value{
A list with
\itemize{
\item{vs, Array of dimensions P,S,S; where P is the number of trials used to train the model and S is the number of stimuli involved in the experiment. Respectively, vs and rs contain the stimulus weights and the stimulus-specific responses.}
\item{alphas, lambdas - Named vectors as above.}
\item{combs, chains - Lists of length P with combined and chained activation values.}
\item{experience, mapping - As passed in arguments; carryover arguments further processing.}
}
}
\description{
Train the RAND model
}
\note{

}
