% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R
\name{results}
\alias{results}
\title{Extract aggregated results from CalmrExperiment}
\usage{
results(object)

results(object)
}
\arguments{
\item{object}{An object of class \clode{\link{CalmrExperiment}}}
}
\value{
A tbl containing models (rows) and model outputs (columns)

A tbl containing models (rows) and model outputs (columns)
}
\description{
Extract aggregated results from CalmrExperiment

Extract aggregated results from CalmrExperiment
}
