% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsa_functions.R, R/class_rsa.R
\name{rsa}
\alias{rsa}
\alias{test,CalmRSA-method}
\alias{plot,CalmRSA-method}
\title{Perform representational similarity analysis on CalmExperiment}
\usage{
rsa(x, comparisons, .test = FALSE, ...)

\S4method{test}{CalmRSA}(object, n_samples = 1000, p = 0.95)

\S4method{plot}{CalmRSA}(x, y, ...)
}
\arguments{
\item{x}{A list of CalmExperiment objects}

\item{comparisons}{A model-named list containing the model
outputs to compare.}

\item{...}{Extra parameters passed to the plot call}

\item{object}{A CalmRSA object}

\item{n_samples}{The number of samples for the permutation test
(default = 1e3)}

\item{p}{The critical threshold level for the permutation test
(default = 0.95)}

\item{test}{Whether to test the RSA via permutation test. Default = FALSE.}
}
\value{
A CalmRSA object

A CalmRSA object with the test results
}
\description{
Perform representational similarity analysis on CalmExperiment

Test CalmRSA object via permutation test
}
\note{
The object returned by this function
can be later tested via its own \code{test} method.
}
\examples{
# Comparing the associations in three models
exp <- data.frame(
  Group = c("A", "B"),
  P1 = c("2(A)>(US)/1B>(US)", "1(A)>(US)/2B>(US)"),
  R1 = TRUE
)
exp <- parse_design(exp)
models <- c("HD2022", "RW1972", "PKH1982")
parameters <- sapply(models, get_parameters, design = exp)
exp_res <- compare_models(exp,
  models = models,
  parameters = parameters
)
comparisons <- list(
  "HD2022" = c("vs"),
  "RW1972" = c("vs"),
  "PKH1982" = c("eivs")
)
res <- rsa(exp_res, comparisons = comparisons)
test(res, n_samples = 100)
}
