% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_experiment.R, R/class_rsa.R
\name{rsa,CalmrExperiment-method}
\alias{rsa,CalmrExperiment-method}
\alias{test,CalmrRSA-method}
\alias{plot,CalmrRSA-method}
\title{Perform representational similarity analysis on CalmrExperiment}
\usage{
\S4method{rsa}{CalmrExperiment}(x, comparisons, test = FALSE, ...)

\S4method{test}{CalmrRSA}(object, n_samples = 1000, p = 0.95)

\S4method{plot}{CalmrRSA}(x, y, ...)
}
\arguments{
\item{x}{A tbl of m by o (models by outputs) with aggregated results.}

\item{comparisons}{A model-named list containing the model
outputs to compare.}

\item{test}{Whether to test the RSA via permutation test. Default = FALSE.}

\item{...}{Additional parameters passed to \code{stats::dist}
and \code{stats::cor}}

\item{object}{The CalmrRSA object}

\item{n_samples}{The number of samples for the permutation test
(default = 1e3)}

\item{p}{The critical threshold level for the permutation test
(default = 0.95)}
}
\value{
A CalmrRSA object

A CalmrRSA object with the test results
}
\description{
Perform representational similarity analysis on CalmrExperiment

Test CalmrRSA object via permutation test
}
\note{
The object returned by this function
can be later tested via its own \code{test} method.
}
\examples{
# Comparing the associations in three models
exp <- data.frame(
  Group = c("A", "B"),
  P1 = c("2(A)>(US)/1B>(US)", "1(A)>(US)/2B>(US)"),
  R1 = TRUE
)
exp <- parse_design(exp)
models <- c("HD2022", "RW1972", "PKH1982")
parameters <- sapply(models, get_parameters, design = exp)
options <- get_exp_opts()
exp_res <- compare_models(exp,
  models = models,
  parameters = parameters, options = options
)
comparisons <- list(
  "HD2022" = c("vs"),
  "RW1972" = c("vs"),
  "PKH1982" = c("eivs")
)
res <- rsa(exp_res, comparisons = comparisons)
test(res, n_samples = 100)
}
