% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rsa.R
\name{test,CalmrRSA-method}
\alias{test,CalmrRSA-method}
\alias{plot,CalmrRSA-method}
\title{Test CalmrRSA object via permutation test}
\usage{
\S4method{test}{CalmrRSA}(object, n_samples = 1000, p = 0.95)

\S4method{plot}{CalmrRSA}(x, y, ...)
}
\arguments{
\item{object}{A CalmrRSA object}

\item{n_samples}{The number of samples for the permutation test
(default = 1e3)}

\item{p}{The critical threshold level for the permutation test
(default = 0.95)}

\item{...}{Extra parameters passed to the plot call}
}
\value{
A CalmrRSA object with the test results
}
\description{
Test CalmrRSA object via permutation test
}
