% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA.R
\name{RSA}
\alias{RSA}
\alias{RSATest}
\title{Perform representational similarity analysis on model comparison}
\usage{
RSA(
  comparison,
  trials = "all",
  dist_method = "euclidean",
  corr_method = "spearman",
  ...
)

RSATest(comparison, n_samples = 1000, p = 0.95, ...)
}
\arguments{
\item{comparison}{A CalmrComparison object, as returned by compare_models}

\item{trials}{A character specifying what trials to use for the comparison. One of \code{c("all", "last")}. Default is "all". See details.}

\item{dist_method}{A character specifying the distance method to be used in \code{dist}. Default is "euclidean".}

\item{corr_method}{A character specifying the correlation method to be used in \code{cor}. Default is "spearman".}

\item{...}{Additional parameters passed to \code{dist} and \code{cor} (in the case of RSA) or \code{RSA} (in the case of RSATest).}

\item{n_samples}{An integer specifying the number of samples for the permutation test. Default is 1e3.}

\item{p}{A float between 0 and 1 specifying the critical value of the null distribution to use for significance testing. Default is 0.95.}
}
\value{
A CalmrRSA or CalmrRSATest object.
}
\description{
Perform representational similarity analysis on model comparison
}
\details{
{
The \code{trials} argument specifyies which trials to use in the calculation of distance matrices; "all" means every trial will be used. "last" means only the last trial of each unique trial type will be used.}
}
