% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Run experiment}
\usage{
run_experiment(
  design,
  model = NULL,
  param_df = NULL,
  options = NULL,
  parse = TRUE
)
}
\arguments{
\item{design}{A data.frame of dimensions G,2*P+1; where G is the number of groups and P is the number of phases.}

\item{model}{A string specifying the model name. One of \code{supported_models()}}

\item{param_df}{A data.frame of dimensions N,2; where N is the number of stimuli in the experimental design.}

\item{options}{A list of options, as returned by \code{get_exp_opts}.}

\item{parse}{A logical specifying whether the results should be parsed via \code{parse_experiment_results}. Default = TRUE.}
}
\value{
A list with parsed results or a tibble with raw results
}
\description{
Runs an experiment with minimal parameters.
}
\seealso{
\code{\link{get_exp_opts}}, \code{\link{parse_experiment_results}}
}
