% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_experiment.R
\name{run_experiment}
\alias{run_experiment}
\title{Run experiment}
\usage{
run_experiment(x, parse = TRUE, aggregate = TRUE, ...)
}
\arguments{
\item{x}{A CalmrExperiment or design data.frame}

\item{parse}{A logical specifying whether the raw results
should be parsed. Default = TRUE.}

\item{aggregate}{A logical specifying whether the parsed results
should be aggregated. Default = TRUE.}

\item{...}{Arguments passed to other functions}
}
\value{
A CalmrExperiment with results.
}
\description{
Runs an experiment with minimal parameters.
}
\examples{
# Using a data.frame only (throws warning)
df <- get_design("relative_validity")
run_experiment(df, model = "RW1972")

# Using custom parameters
df <- get_design("relative_validity")
pars <- get_parameters(df, model = "HD2022")
pars$alphas["US"] <- 0.6
run_experiment(df, parameters = pars, model = "HD2022")

# Using CalmrExperiment, for more iterations
df <- get_design("blocking")
pars <- get_parameters(df, model = "SM2007")
exper <- make_experiment(df,
  parameters = pars, model = "SM2007",
  options = get_exp_opts(iterations = 4)
)
run_experiment(exper)
}
\seealso{
\code{\link{get_exp_opts}}, \code{\link{parse_experiment_results}}
}
