% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_heidi.R
\name{run_heidi}
\alias{run_heidi}
\title{Run the HeiDI model}
\usage{
run_heidi(args, parse = TRUE)
}
\arguments{
\item{args}{A tibble as returned by \code{make_heidi_args}.}

\item{parse}{A logical specifying whether the results should be parsed. Default = TRUE.}
}
\value{
A list with parsed results or a tibble with raw results
}
\description{
Runs the model under specific arguments.
}
\seealso{
\code{\link{make_heidi_args}}
}
