% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model.R
\name{run_model}
\alias{run_model}
\title{Run a model}
\usage{
run_model(args, parse, ...)
}
\arguments{
\item{args}{A tibble as returned by \code{make_model_args}.}

\item{parse}{A logical specifying whether the results should be parsed.}
}
\value{
A CalmrExperiment object.
}
\description{
Runs a model with specific arguments.
}
\seealso{
\code{\link{make_model_args}}
}
