% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SM2007.R
\name{SM2007}
\alias{SM2007}
\title{Train the SM2007 model}
\usage{
SM2007(
  v = NULL,
  o = NULL,
  parameters,
  experience,
  mapping,
  debug = FALSE,
  comparator_func = .witnauer_comparator_proc,
  ...
)
}
\arguments{
\item{v}{(optional) A named matrix of dimensions S,S;
where S is the number of stimuli.}

\item{o}{(optional) A named matrix of dimensions S,S,S
(the operator switches).}

\item{parameters}{A list containing the model parameters,
as returned by get_parameters().}

\item{experience}{A data.frame specifying trials as rows,
as returned by \code{make_experiment}}

\item{mapping}{A named list specifying trial and stimulus mapping,
as returned by \code{make_experiment}}

\item{debug}{A logical specifying whether to print
information for the comparison process. Defaults to FALSE.}

\item{comparator_func}{A function to be used in the comparison
process. Either \code{.witnauer_comparator_func}or \code{.comparator_func}.}

\item{...}{Additional named arguments}
}
\value{
A list with raw results
}
\description{
Train the SM2007 model
}
\note{
Correct usage of the v and o parameters requires the
matrices to adhere to a specific format.
For v, entry i,j represents the association from
stimulus i to stimulus j.
For o, entry i,k,j represents the operator switch
between stimulus i and comparator k with respect to stimulus j.
}
