% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment_class.R
\name{summary,CalmrExperiment-method}
\alias{summary,CalmrExperiment-method}
\title{Summarise CalmrExperiment}
\usage{
\S4method{summary}{CalmrExperiment}(object, ...)
}
\arguments{
\item{object}{An object of class \code{\link{CalmrExperiment}}.}

\item{...}{Additional parameters passed to the summary function.}
}
\value{
A
}
\description{
Shows a summary for a CalmrExperiment
}
