% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pav_heidi.R
\name{train_pav_heidi}
\alias{train_pav_heidi}
\title{Trains a Pavlovian HeiDI}
\usage{
train_pav_heidi(
  sals,
  V,
  tps,
  trial_pre_func,
  trial_post_func,
  trial_pre_nomi,
  trial_post_nomi,
  nomi_func_map,
  trial_names,
  phase,
  block_size,
  is_test
)
}
\arguments{
\item{sals}{A named vector with stimulus saliences.}

\item{V}{A named matrix of dimensions S,S; where S is the number of stimuli.}

\item{tps}{A vector of trial pointers for training, as a function of trials.}

\item{trial_pre_func}{A list of length T, with character vectors specifying the functional stimuli involved in the expectation part of each trial. T is the number of unique trials in the experiment.}

\item{trial_post_func}{As above, but for the correction part of the trial.}

\item{trial_pre_nomi}{A list of length T, with character vectors specifying the nominal stimuli involved in the expectation part of each trial.}

\item{trial_post_nomi}{As above, but for the correction part of the trial.}

\item{nomi_func_map}{A data.frame with the mappings between nominal and functional stimuli}

\item{trial_names}{A character vector of length T with the names of the trials}

\item{phase}{A character vector of length T with the names of the phases}

\item{block_size}{A integer vector of length T specifying the block size per trial}

\item{is_test}{A logical vector specifying whether the trial should result in learning (update V). If an element is TRUE, no update occurs.}
}
\value{
A list with
\itemize{
\item{vs, rs - Arrays of dimensions P,S,S; where P is the number of trials used to train the model and S is the number of stimuli involved in the experiment. Respectively, vs and rs contain the stimulus weights and the stimulus-specific responses.}
\item{combs, chains - Lists of length P with combined and chained activation values.}
\item{tps, trial_pre_func, trial_post_func, trial_pre_nomi, trial_post_nomi, nomi_func_map, trial_names, phase, block_size, is_test - Carryover arguments further processing.}
}
}
\description{
Trains a Pavlovian HeiDI
}
\note{
The array V contains the associations for all stimuli involved in the experiment. Entry i,j specifies the associative strength between stimulus i to stimulus j. Entry j,i specifies the opposite direction.
}
