% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trial_parser.R
\name{trial_parser}
\alias{trial_parser}
\title{Parses a string stipulating trials}
\usage{
trial_parser(str)
}
\arguments{
\item{str}{A string}
}
\value{
A list with the number of trials, the trials, and the stimulus names
}
\description{
Parses a string stipulating trials
}
\note{
Supports complex stimuli in parentheses (e.g., "5A(US)")
Supports multiple versions of stimuli (e.g., "3(A_a)(US_a)/3(A_b)(US_b)")
Supports probe trials, followed by a hash (e.g., "3X#")
}
\examples{
trial_parser("30XAU/30(YA)A")
trial_parser("1X#/2X(US)#")
trial_parser("1X(US_a)/1Y(US_b)")
}
